MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for Spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for Sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object

A spatial regression model object fitted by maximum likelihood with spautolm

mcmc

The number of MCMC iterations after burnin

verbose

default NULL, use global option value; if TRUE, reports progress

...

Arguments passed through

burnin

The number of burn-in iterations for the sampler

scale

a positive scale parameter

listw, listw2

listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control

list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

Author

Roger Bivand Roger.Bivand@nhh.no

See also

Examples

require("sf", quietly=TRUE)
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
#> [1] TRUE
#require("spdep", quietly=TRUE)
nyadjlw <- spdep::mat2listw(nyadjmat, as.character(nydata$AREAKEY))
listw_NY <- spdep::nb2listw(nyadjlw$neighbours, style="B")
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017209 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 
res <- MCMCsamp(esar1f, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04452 0.01714 0.0005421       0.002169
#> (Intercept) -0.63992 0.19291 0.0061005       0.027279
#> PEXPOSURE    0.08123 0.04583 0.0014491       0.005692
#> PCTAGE65P    3.73439 0.52232 0.0165171       0.059784
#> PCTOWNHOME  -0.40228 0.20466 0.0064720       0.028307
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.009424  0.03451  0.04492  0.05674  0.07778
#> (Intercept) -1.065706 -0.77641 -0.62900 -0.50557 -0.24289
#> PEXPOSURE   -0.013777  0.05427  0.08662  0.10736  0.16850
#> PCTAGE65P    2.723922  3.35581  3.72856  4.09109  4.88304
#> PCTOWNHOME  -0.787942 -0.52515 -0.40396 -0.27885  0.01044
#> 
# \dontrun{
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01427 0.01656 0.0002342      0.0009576
#> (Intercept) -0.80700 0.16169 0.0022866      0.0108559
#> PEXPOSURE    0.08073 0.03097 0.0004380      0.0018595
#> PCTAGE65P    3.84765 0.61366 0.0086785      0.0360672
#> PCTOWNHOME  -0.37980 0.17020 0.0024071      0.0103135
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.01657  0.00300  0.01386  0.02579  0.04697
#> (Intercept) -1.12218 -0.91749 -0.80536 -0.70427 -0.48217
#> PEXPOSURE    0.01902  0.05936  0.08129  0.10136  0.13953
#> PCTAGE65P    2.60531  3.43050  3.82234  4.26808  4.98684
#> PCTOWNHOME  -0.71930 -0.48531 -0.38309 -0.26722 -0.04716
#> 
ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 
res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08847 0.02911 0.0004116       0.001653
#> (Intercept) -0.68402 0.20325 0.0028743       0.012173
#> PEXPOSURE    0.08921 0.05083 0.0007189       0.003198
#> PCTAGE65P    3.69190 0.67495 0.0095452       0.040188
#> PCTOWNHOME  -0.36381 0.22788 0.0032227       0.014146
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%     75%    97.5%
#> lambda       0.02509  0.06949  0.09114  0.1102  0.13543
#> (Intercept) -1.08344 -0.82879 -0.67742 -0.5498 -0.29054
#> PEXPOSURE   -0.00709  0.05472  0.08864  0.1185  0.20415
#> PCTAGE65P    2.42182  3.23361  3.66133  4.1273  5.02996
#> PCTOWNHOME  -0.81618 -0.51525 -0.37261 -0.2076  0.07076
#> 
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01453 0.01655 0.0002340      0.0009397
#> (Intercept) -0.81370 0.16045 0.0022691      0.0100733
#> PEXPOSURE    0.08301 0.03329 0.0004707      0.0021269
#> PCTAGE65P    3.80312 0.61044 0.0086329      0.0361088
#> PCTOWNHOME  -0.36508 0.16897 0.0023896      0.0099003
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01831  0.003536  0.01518  0.02585  0.04542
#> (Intercept) -1.12644 -0.921052 -0.81114 -0.70248 -0.49510
#> PEXPOSURE    0.02006  0.060386  0.08222  0.10454  0.14938
#> PCTAGE65P    2.62810  3.392689  3.80821  4.23210  4.97774
#> PCTOWNHOME  -0.68794 -0.479899 -0.37006 -0.24582 -0.01695
#> 
ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 
res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.03963 0.03997 0.0005652       0.002601
#> (Intercept) -0.83720 0.17086 0.0024164       0.010918
#> PEXPOSURE    0.09102 0.03314 0.0004687       0.001994
#> PCTAGE65P    3.72954 0.61839 0.0087453       0.036753
#> PCTOWNHOME  -0.34971 0.17640 0.0024947       0.010774
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%     75%    97.5%
#> lambda      -0.04452  0.01383  0.04145  0.0664  0.11802
#> (Intercept) -1.21583 -0.94042 -0.82729 -0.7183 -0.53054
#> PEXPOSURE    0.02931  0.06925  0.08970  0.1127  0.16140
#> PCTAGE65P    2.50950  3.33565  3.74064  4.1619  4.90395
#> PCTOWNHOME  -0.66984 -0.47682 -0.35855 -0.2340  0.04174
#> 
# }
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04090 0.01698 0.0005371       0.002400
#> (Intercept) -0.64408 0.15867 0.0050177       0.019082
#> PEXPOSURE    0.07739 0.04134 0.0013072       0.005386
#> PCTAGE65P    3.79901 0.58114 0.0183772       0.074409
#> PCTOWNHOME  -0.42161 0.19745 0.0062439       0.026380
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.005269  0.02851  0.04092  0.05306  0.07233
#> (Intercept) -0.933261 -0.75817 -0.65712 -0.52703 -0.33683
#> PEXPOSURE   -0.011686  0.05227  0.07668  0.10420  0.15994
#> PCTAGE65P    2.749133  3.42605  3.74635  4.18424  4.98368
#> PCTOWNHOME  -0.822524 -0.53737 -0.40482 -0.29829 -0.04847
#> 
# \dontrun{
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01367 0.01604 0.0002269      0.0008963
#> (Intercept) -0.81297 0.14285 0.0020202      0.0078212
#> PEXPOSURE    0.08365 0.03062 0.0004330      0.0019426
#> PCTAGE65P    3.80898 0.60246 0.0085200      0.0350542
#> PCTOWNHOME  -0.36646 0.15773 0.0022307      0.0088482
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01666  0.002323  0.01381  0.02477  0.04477
#> (Intercept) -1.10730 -0.904518 -0.81194 -0.71410 -0.54227
#> PEXPOSURE    0.02648  0.061513  0.08339  0.10545  0.14415
#> PCTAGE65P    2.56607  3.430172  3.81218  4.19964  4.97072
#> PCTOWNHOME  -0.66993 -0.471535 -0.37171 -0.26092 -0.04913
#> 
esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 
res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02997 0.01680 0.0002376       0.001448
#> (Intercept)     -1.11681 0.24098 0.0034080       0.018298
#> PEXPOSURE        0.22819 0.07748 0.0010957       0.005865
#> PCTAGE65P        3.37575 0.64844 0.0091703       0.052645
#> PCTOWNHOME       0.02148 0.24163 0.0034172       0.019352
#> lag.(Intercept)  0.12102 0.05855 0.0008281       0.004638
#> lag.PEXPOSURE   -0.03633 0.01518 0.0002147       0.001079
#> lag.PCTAGE65P    0.25353 0.23853 0.0033734       0.020306
#> lag.PCTOWNHOME  -0.15022 0.06144 0.0008690       0.004949
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> lambda           0.001075  0.01700  0.02908  0.04135  0.064775
#> (Intercept)     -1.587953 -1.28475 -1.11591 -0.96311 -0.617616
#> PEXPOSURE        0.075072  0.17158  0.22555  0.28248  0.379416
#> PCTAGE65P        2.144838  2.96654  3.40236  3.80900  4.695992
#> PCTOWNHOME      -0.449133 -0.14798  0.03525  0.18560  0.495676
#> lag.(Intercept)  0.002054  0.07909  0.11953  0.16147  0.234027
#> lag.PEXPOSURE   -0.064609 -0.04664 -0.03632 -0.02639 -0.007153
#> lag.PCTAGE65P   -0.261274  0.10051  0.25968  0.42440  0.666460
#> lag.PCTOWNHOME  -0.271058 -0.19043 -0.14791 -0.10881 -0.031695
#> 
lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 
res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03805 0.01371 0.0001939      0.0007109
#> (Intercept) -0.51602 0.15517 0.0021944      0.0088771
#> PEXPOSURE    0.04546 0.03199 0.0004524      0.0016532
#> PCTAGE65P    3.66422 0.59646 0.0084353      0.0336477
#> PCTOWNHOME  -0.40866 0.17293 0.0024456      0.0101148
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> rho          0.00978  0.02899  0.03850  0.04725  0.06314
#> (Intercept) -0.83111 -0.61597 -0.51660 -0.41321 -0.20402
#> PEXPOSURE   -0.01883  0.02451  0.04566  0.06612  0.10932
#> PCTAGE65P    2.46650  3.23821  3.67313  4.08788  4.77441
#> PCTOWNHOME  -0.73752 -0.53103 -0.41243 -0.29823 -0.07508
#> 
lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 
res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02343 0.01721 0.0002433       0.001264
#> (Intercept)     -1.14420 0.25097 0.0035493       0.018903
#> PEXPOSURE        0.21295 0.08676 0.0012269       0.006956
#> PCTAGE65P        3.45520 0.67283 0.0095153       0.050549
#> PCTOWNHOME       0.09540 0.25445 0.0035984       0.019452
#> lag.(Intercept)  0.13681 0.05476 0.0007745       0.003919
#> lag.PEXPOSURE   -0.03454 0.01667 0.0002358       0.001282
#> lag.PCTAGE65P    0.15250 0.23263 0.0032899       0.017657
#> lag.PCTOWNHOME  -0.15052 0.05929 0.0008385       0.004441
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%      25%      50%      75%     97.5%
#> rho             -0.01021  0.01123  0.02372  0.03566  0.055933
#> (Intercept)     -1.65804 -1.30366 -1.12556 -0.96823 -0.667178
#> PEXPOSURE        0.05192  0.14749  0.21585  0.26994  0.390874
#> PCTAGE65P        2.05107  2.96840  3.45858  3.95580  4.728122
#> PCTOWNHOME      -0.39229 -0.08432  0.09188  0.27309  0.584920
#> lag.(Intercept)  0.03739  0.09877  0.13902  0.17159  0.246457
#> lag.PEXPOSURE   -0.06738 -0.04583 -0.03458 -0.02307 -0.002202
#> lag.PCTAGE65P   -0.29775 -0.01050  0.14361  0.31995  0.607184
#> lag.PCTOWNHOME  -0.26866 -0.18885 -0.15289 -0.10711 -0.032731
#> 
ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho         -0.02299 0.07704 0.0010896        0.02224
#> lambda       0.04617 0.07769 0.0010988        0.01920
#> (Intercept) -0.72343 0.30240 0.0042766        0.05976
#> PEXPOSURE    0.09297 0.06811 0.0009632        0.01362
#> PCTAGE65P    3.38473 0.61518 0.0086999        0.04126
#> PCTOWNHOME  -0.24532 0.23038 0.0032581        0.02571
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> rho         -0.139314 -0.09171 -0.03549  0.05275  0.09903
#> lambda      -0.101969 -0.02427  0.07760  0.11500  0.13610
#> (Intercept) -1.322627 -0.96417 -0.69707 -0.47385 -0.25428
#> PEXPOSURE   -0.003173  0.03862  0.07807  0.14471  0.23922
#> PCTAGE65P    2.113791  2.98783  3.40026  3.81625  4.56204
#> PCTOWNHOME  -0.671554 -0.41180 -0.26174 -0.09882  0.27737
#> 
ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho             -0.04461 0.07857 0.0011112       0.025966
#> lambda           0.05380 0.07029 0.0009941       0.020643
#> (Intercept)     -1.12556 0.23162 0.0032756       0.016490
#> PEXPOSURE        0.21456 0.07595 0.0010741       0.005910
#> PCTAGE65P        3.40318 0.67222 0.0095067       0.056196
#> PCTOWNHOME       0.05142 0.26163 0.0037000       0.021392
#> lag.(Intercept)  0.05553 0.10771 0.0015232       0.023925
#> lag.PEXPOSURE   -0.03144 0.02347 0.0003318       0.005415
#> lag.PCTAGE65P    0.44560 0.35575 0.0050310       0.060235
#> lag.PCTOWNHOME  -0.11549 0.06817 0.0009640       0.006865
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%       25%      50%      75%    97.5%
#> rho             -0.16653 -0.110786 -0.05525  0.02552  0.09707
#> lambda          -0.09757  0.003543  0.07765  0.11297  0.13338
#> (Intercept)     -1.55705 -1.272641 -1.12264 -0.99044 -0.64994
#> PEXPOSURE        0.07981  0.165979  0.21222  0.25756  0.36892
#> PCTAGE65P        1.97509  3.023208  3.40603  3.84495  4.72195
#> PCTOWNHOME      -0.48433 -0.102685  0.06265  0.21867  0.58297
#> lag.(Intercept) -0.16764 -0.014227  0.08311  0.14225  0.21388
#> lag.PEXPOSURE   -0.06983 -0.047663 -0.03488 -0.01672  0.01412
#> lag.PCTAGE65P   -0.20380  0.193496  0.44589  0.71810  1.13271
#> lag.PCTOWNHOME  -0.24146 -0.169111 -0.11455 -0.06601  0.02117
#> 
# }