MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for Spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for Sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object

A spatial regression model object fitted by maximum likelihood with spautolm

mcmc

The number of MCMC iterations after burnin

verbose

default NULL, use global option value; if TRUE, reports progress

...

Arguments passed through

burnin

The number of burn-in iterations for the sampler

scale

a positive scale parameter

listw, listw2

listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control

list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

Author

Roger Bivand Roger.Bivand@nhh.no

See also

Examples

require("sf", quietly=TRUE)
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
#> [1] TRUE
#require("spdep", quietly=TRUE)
nyadjlw <- spdep::mat2listw(nyadjmat, as.character(nydata$AREAKEY))
listw_NY <- spdep::nb2listw(nyadjlw$neighbours, style="B")
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017209 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 
res <- MCMCsamp(esar1f, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04894 0.01922 0.0006079       0.003065
#> (Intercept) -0.71507 0.21985 0.0069523       0.034579
#> PEXPOSURE    0.09203 0.05264 0.0016647       0.008383
#> PCTAGE65P    3.84521 0.66189 0.0209309       0.102778
#> PCTOWNHOME  -0.32388 0.23465 0.0074202       0.034193
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.008785  0.03746  0.04728  0.06361  0.09229
#> (Intercept) -1.361708 -0.83196 -0.68848 -0.57364 -0.37457
#> PEXPOSURE    0.001016  0.05934  0.08560  0.11773  0.22585
#> PCTAGE65P    2.565507  3.37536  3.83786  4.24711  5.34785
#> PCTOWNHOME  -0.747795 -0.49952 -0.34027 -0.15493  0.22261
#> 
# \dontrun{
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01538 0.01669 0.0002360      0.0009728
#> (Intercept) -0.81506 0.14977 0.0021180      0.0088564
#> PEXPOSURE    0.08407 0.02882 0.0004075      0.0016909
#> PCTAGE65P    3.79024 0.59110 0.0083594      0.0342490
#> PCTOWNHOME  -0.36061 0.15705 0.0022210      0.0090039
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01771  0.004466  0.01543  0.02693  0.04862
#> (Intercept) -1.10147 -0.915936 -0.81454 -0.71937 -0.51654
#> PEXPOSURE    0.02717  0.064254  0.08369  0.10428  0.14159
#> PCTAGE65P    2.61169  3.399023  3.78032  4.18672  4.92309
#> PCTOWNHOME  -0.67264 -0.465559 -0.35869 -0.25336 -0.06609
#> 
ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 
res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08758 0.03148 0.0004451       0.001954
#> (Intercept) -0.67629 0.20192 0.0028556       0.012083
#> PEXPOSURE    0.08603 0.04810 0.0006802       0.002847
#> PCTAGE65P    3.66509 0.59815 0.0084591       0.031845
#> PCTOWNHOME  -0.34996 0.22804 0.0032250       0.014079
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%     75%    97.5%
#> lambda       0.021924  0.06759  0.09203  0.1106  0.13680
#> (Intercept) -1.076177 -0.80961 -0.67331 -0.5460 -0.28612
#> PEXPOSURE   -0.001706  0.05388  0.08326  0.1155  0.18770
#> PCTAGE65P    2.473348  3.24278  3.65480  4.0901  4.78981
#> PCTOWNHOME  -0.780852 -0.50425 -0.36278 -0.1926  0.08878
#> 
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01211 0.01694 0.0002396      0.0009508
#> (Intercept) -0.80954 0.15485 0.0021899      0.0094683
#> PEXPOSURE    0.08231 0.03025 0.0004278      0.0017934
#> PCTAGE65P    3.84838 0.60238 0.0085189      0.0362146
#> PCTOWNHOME  -0.37351 0.17579 0.0024860      0.0113555
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.02077  0.000556  0.01237  0.02363  0.04529
#> (Intercept) -1.12271 -0.908607 -0.80359 -0.70730 -0.52055
#> PEXPOSURE    0.02775  0.062318  0.08002  0.10082  0.14701
#> PCTAGE65P    2.62003  3.448681  3.83388  4.28111  5.01553
#> PCTOWNHOME  -0.69674 -0.490475 -0.37639 -0.25990 -0.02412
#> 
ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 
res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04070 0.04026 0.0005694       0.002328
#> (Intercept) -0.83166 0.15742 0.0022262       0.008973
#> PEXPOSURE    0.09004 0.03383 0.0004784       0.001990
#> PCTAGE65P    3.67368 0.57880 0.0081855       0.031793
#> PCTOWNHOME  -0.35141 0.17501 0.0024750       0.010472
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%     97.5%
#> lambda      -0.04123  0.01354  0.04161  0.07003  0.114929
#> (Intercept) -1.14459 -0.93906 -0.82465 -0.72544 -0.540834
#> PEXPOSURE    0.02626  0.06545  0.09158  0.11279  0.158922
#> PCTAGE65P    2.54741  3.30037  3.67397  4.07741  4.745613
#> PCTOWNHOME  -0.67388 -0.47296 -0.34776 -0.23393  0.002485
#> 
# }
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04823 0.01738 0.0005495       0.002339
#> (Intercept) -0.66033 0.19544 0.0061805       0.026758
#> PEXPOSURE    0.09008 0.04877 0.0015422       0.007009
#> PCTAGE65P    3.72183 0.65559 0.0207314       0.117996
#> PCTOWNHOME  -0.37964 0.20789 0.0065740       0.027629
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.015303  0.03545  0.04912  0.05996  0.08169
#> (Intercept) -1.099620 -0.76653 -0.65046 -0.54927 -0.22351
#> PEXPOSURE    0.005719  0.05150  0.09355  0.11868  0.18684
#> PCTAGE65P    2.410499  3.31891  3.67177  4.10652  5.14582
#> PCTOWNHOME  -0.754253 -0.52750 -0.40871 -0.23750  0.03468
#> 
# \dontrun{
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01265 0.01668 0.0002359      0.0009203
#> (Intercept) -0.80611 0.14658 0.0020729      0.0088682
#> PEXPOSURE    0.08402 0.03056 0.0004322      0.0019017
#> PCTAGE65P    3.79710 0.59889 0.0084696      0.0347044
#> PCTOWNHOME  -0.37150 0.15171 0.0021455      0.0084474
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%        25%      50%      75%    97.5%
#> lambda      -0.02021  0.0008258  0.01211  0.02423  0.04343
#> (Intercept) -1.11214 -0.9014124 -0.80991 -0.70885 -0.51529
#> PEXPOSURE    0.02554  0.0620309  0.08343  0.10520  0.14517
#> PCTAGE65P    2.67047  3.3918411  3.80467  4.20719  4.95639
#> PCTOWNHOME  -0.67479 -0.4721760 -0.37251 -0.26667 -0.07344
#> 
esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 
res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02451 0.01778 0.0002515       0.001451
#> (Intercept)     -1.12645 0.23914 0.0033820       0.017653
#> PEXPOSURE        0.22167 0.07626 0.0010785       0.005904
#> PCTAGE65P        3.35565 0.67647 0.0095668       0.053330
#> PCTOWNHOME       0.04299 0.24694 0.0034922       0.015726
#> lag.(Intercept)  0.12417 0.05754 0.0008137       0.004285
#> lag.PEXPOSURE   -0.03529 0.01664 0.0002353       0.001386
#> lag.PCTAGE65P    0.25954 0.24287 0.0034347       0.019913
#> lag.PCTOWNHOME  -0.15682 0.06002 0.0008488       0.004661
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%      25%      50%      75%    97.5%
#> lambda          -0.01162  0.01361  0.02480  0.03736  0.05743
#> (Intercept)     -1.61032 -1.28447 -1.13914 -0.95612 -0.65400
#> PEXPOSURE        0.08005  0.16977  0.21896  0.26965  0.37978
#> PCTAGE65P        2.01987  2.89826  3.38143  3.81627  4.71393
#> PCTOWNHOME      -0.47631 -0.11378  0.04167  0.19252  0.54606
#> lag.(Intercept)  0.01500  0.08283  0.12142  0.16248  0.23227
#> lag.PEXPOSURE   -0.07006 -0.04609 -0.03431 -0.02278 -0.00732
#> lag.PCTAGE65P   -0.25735  0.10328  0.26439  0.43009  0.70925
#> lag.PCTOWNHOME  -0.27684 -0.19656 -0.15976 -0.11702 -0.03737
#> 
lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 
res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03946 0.01418 0.0002005      0.0007803
#> (Intercept) -0.51258 0.14966 0.0021165      0.0080478
#> PEXPOSURE    0.04890 0.03302 0.0004670      0.0017684
#> PCTAGE65P    3.62986 0.61248 0.0086618      0.0355727
#> PCTOWNHOME  -0.42408 0.16624 0.0023510      0.0092138
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> rho          0.01140  0.02995  0.03941  0.04973  0.06551
#> (Intercept) -0.80617 -0.60676 -0.51128 -0.41321 -0.21307
#> PEXPOSURE   -0.01705  0.02651  0.04830  0.07172  0.11369
#> PCTAGE65P    2.41210  3.24537  3.63907  4.04993  4.77435
#> PCTOWNHOME  -0.77029 -0.53348 -0.41998 -0.30922 -0.10850
#> 
lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 
res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02527 0.01634 0.0002310       0.001314
#> (Intercept)     -1.13709 0.24434 0.0034556       0.019992
#> PEXPOSURE        0.21677 0.08442 0.0011939       0.007046
#> PCTAGE65P        3.36288 0.64754 0.0091575       0.049777
#> PCTOWNHOME       0.05703 0.25070 0.0035454       0.020618
#> lag.(Intercept)  0.13167 0.05739 0.0008116       0.004481
#> lag.PEXPOSURE   -0.03543 0.01670 0.0002361       0.001476
#> lag.PCTAGE65P    0.16907 0.21796 0.0030824       0.016462
#> lag.PCTOWNHOME  -0.13617 0.06462 0.0009138       0.005647
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> rho             -0.006074  0.01411  0.02487  0.03579  0.055643
#> (Intercept)     -1.606690 -1.29409 -1.15633 -0.98727 -0.635462
#> PEXPOSURE        0.042473  0.16715  0.21792  0.26670  0.391207
#> PCTAGE65P        2.196269  2.93844  3.34183  3.74740  4.720069
#> PCTOWNHOME      -0.436396 -0.09964  0.04666  0.23013  0.538902
#> lag.(Intercept)  0.020450  0.09538  0.13112  0.16852  0.252210
#> lag.PEXPOSURE   -0.073721 -0.04438 -0.03484 -0.02629 -0.002931
#> lag.PCTAGE65P   -0.251624  0.01551  0.16519  0.31825  0.606798
#> lag.PCTOWNHOME  -0.264790 -0.17990 -0.13576 -0.09126 -0.005181
#> 
ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> rho         -0.03159 0.07162 0.001013        0.01799
#> lambda       0.05772 0.07141 0.001010        0.01551
#> (Intercept) -0.76777 0.30473 0.004309        0.04779
#> PEXPOSURE    0.11502 0.08341 0.001180        0.01585
#> PCTAGE65P    3.35015 0.68020 0.009619        0.05540
#> PCTOWNHOME  -0.27757 0.23050 0.003260        0.02380
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%       25%      50%      75%    97.5%
#> rho         -0.137206 -0.094824 -0.04459  0.03613  0.09388
#> lambda      -0.098512  0.002571  0.08537  0.11740  0.13785
#> (Intercept) -1.350100 -0.998114 -0.76066 -0.52182 -0.24684
#> PEXPOSURE   -0.007896  0.045788  0.10507  0.17443  0.28731
#> PCTAGE65P    2.041453  2.876719  3.36324  3.83423  4.68874
#> PCTOWNHOME  -0.691209 -0.447091 -0.28418 -0.12099  0.20805
#> 
ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                      Mean      SD  Naive SE Time-series SE
#> rho             -0.033944 0.06860 0.0009701       0.017036
#> lambda           0.048247 0.06337 0.0008962       0.014028
#> (Intercept)     -1.067612 0.26620 0.0037647       0.020550
#> PEXPOSURE        0.215408 0.07618 0.0010774       0.005554
#> PCTAGE65P        3.329494 0.77780 0.0109998       0.067453
#> PCTOWNHOME      -0.009999 0.28320 0.0040051       0.022161
#> lag.(Intercept)  0.103733 0.07371 0.0010425       0.008968
#> lag.PEXPOSURE   -0.028843 0.02005 0.0002836       0.002507
#> lag.PCTAGE65P    0.277482 0.34575 0.0048897       0.044802
#> lag.PCTOWNHOME  -0.165116 0.07940 0.0011229       0.009821
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%       25%       50%      75%    97.5%
#> rho             -0.15139 -0.083763 -0.045766  0.01583  0.09736
#> lambda          -0.09093  0.008208  0.064517  0.09768  0.13301
#> (Intercept)     -1.60394 -1.236613 -1.082727 -0.89258 -0.51793
#> PEXPOSURE        0.06566  0.165337  0.218087  0.26718  0.36813
#> PCTAGE65P        1.66548  2.828330  3.357752  3.90993  4.69973
#> PCTOWNHOME      -0.55717 -0.216775 -0.009207  0.16786  0.57330
#> lag.(Intercept) -0.05099  0.052830  0.104882  0.15503  0.23472
#> lag.PEXPOSURE   -0.06595 -0.042349 -0.029459 -0.01684  0.01356
#> lag.PCTAGE65P   -0.37199  0.031113  0.288762  0.50471  1.01153
#> lag.PCTOWNHOME  -0.31939 -0.214946 -0.165399 -0.10886 -0.01901
#> 
# }