MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for Spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for Sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object
A spatial regression model object fitted by maximum likelihood with spautolm

mcmc
The number of MCMC iterations after burnin

verbose
default NULL, use global option value; if TRUE, reports progress

...
Arguments passed through

burnin
The number of burn-in iterations for the sampler

scale
a positive scale parameter

listw, listw2
listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control
list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

Author

Roger Bivand Roger.Bivand@nhh.no

See also

Examples

require("sf", quietly=TRUE)
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))

#> [1] TRUE

#require("spdep", quietly=TRUE)
nyadjlw <- spdep::mat2listw(nyadjmat, as.character(nydata$AREAKEY))
listw_NY <- spdep::nb2listw(nyadjlw$neighbours, style="B")
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017209 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 

res <- MCMCsamp(esar1f, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04312 0.01738 0.0005496       0.002193
#> (Intercept) -0.64883 0.18254 0.0057724       0.023256
#> PEXPOSURE    0.08483 0.04913 0.0015537       0.006320
#> PCTAGE65P    3.78232 0.58084 0.0183678       0.066337
#> PCTOWNHOME  -0.39612 0.22118 0.0069943       0.030304
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%     97.5%
#> lambda       0.008390  0.03207  0.04215  0.05501  0.076133
#> (Intercept) -1.039676 -0.77587 -0.64419 -0.51195 -0.326557
#> PEXPOSURE   -0.004963  0.05359  0.08142  0.11521  0.202348
#> PCTAGE65P    2.627258  3.39980  3.73455  4.18334  4.962310
#> PCTOWNHOME  -0.819651 -0.56246 -0.39274 -0.22634  0.004321
#> 

# \dontrun{
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01271 0.01597 0.0002258       0.000835
#> (Intercept) -0.79589 0.16204 0.0022916       0.009234
#> PEXPOSURE    0.08135 0.03004 0.0004249       0.001609
#> PCTAGE65P    3.81344 0.61395 0.0086826       0.035900
#> PCTOWNHOME  -0.38876 0.17337 0.0024519       0.010405
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.01950  0.00131  0.01365  0.02322  0.04405
#> (Intercept) -1.11386 -0.90118 -0.79245 -0.68884 -0.47880
#> PEXPOSURE    0.02468  0.06054  0.08122  0.10113  0.13929
#> PCTAGE65P    2.58222  3.40881  3.80753  4.22443  4.99183
#> PCTOWNHOME  -0.74173 -0.49840 -0.38546 -0.27256 -0.05256
#> 

ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 

res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08485 0.03006 0.0004252       0.001841
#> (Intercept) -0.66321 0.21539 0.0030460       0.014590
#> PEXPOSURE    0.08242 0.04990 0.0007057       0.003038
#> PCTAGE65P    3.65268 0.64313 0.0090952       0.039343
#> PCTOWNHOME  -0.35758 0.22527 0.0031859       0.014859
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%     75%   97.5%
#> lambda       0.021082  0.06566  0.08492  0.1070  0.1398
#> (Intercept) -1.173283 -0.78620 -0.63969 -0.5232 -0.2864
#> PEXPOSURE   -0.008505  0.04925  0.07835  0.1125  0.1967
#> PCTAGE65P    2.400072  3.17780  3.65483  4.0905  4.9434
#> PCTOWNHOME  -0.761276 -0.51653 -0.37306 -0.2167  0.1509
#> 

esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 

res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01349 0.01687 0.0002386      0.0009539
#> (Intercept) -0.78310 0.16153 0.0022843      0.0098285
#> PEXPOSURE    0.07981 0.03110 0.0004398      0.0018722
#> PCTAGE65P    3.75766 0.64163 0.0090741      0.0389019
#> PCTOWNHOME  -0.39515 0.17309 0.0024478      0.0098508
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01923  0.001626  0.01392  0.02433  0.04500
#> (Intercept) -1.09907 -0.894216 -0.78133 -0.67539 -0.46521
#> PEXPOSURE    0.02333  0.057110  0.07880  0.10072  0.14261
#> PCTAGE65P    2.45223  3.345132  3.76443  4.20398  5.00203
#> PCTOWNHOME  -0.72512 -0.516227 -0.39431 -0.27794 -0.06093
#> 

ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)

#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 

res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.03533 0.03881 0.0005488       0.002185
#> (Intercept) -0.83029 0.16302 0.0023055       0.010457
#> PEXPOSURE    0.08737 0.03435 0.0004858       0.002162
#> PCTAGE65P    3.75700 0.56221 0.0079509       0.031835
#> PCTOWNHOME  -0.34683 0.17576 0.0024856       0.011264
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.04410  0.009632  0.03641  0.06253  0.10672
#> (Intercept) -1.15205 -0.939630 -0.82682 -0.71785 -0.51312
#> PEXPOSURE    0.02537  0.063595  0.08462  0.10872  0.16356
#> PCTAGE65P    2.64044  3.356670  3.76310  4.13533  4.81561
#> PCTOWNHOME  -0.67554 -0.468450 -0.34818 -0.23379  0.00413
#> 

# }
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(esar0, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04682 0.01895 0.0005993       0.002739
#> (Intercept) -0.65775 0.21896 0.0069243       0.031633
#> PEXPOSURE    0.08543 0.05260 0.0016634       0.007626
#> PCTAGE65P    3.65044 0.57371 0.0181422       0.060983
#> PCTOWNHOME  -0.36116 0.23745 0.0075089       0.036472
#> 
#> 2. Quantiles for each variable:
#> 
#>                   2.5%      25%      50%      75%    97.5%
#> lambda       0.0109828  0.03340  0.04839  0.05958  0.08139
#> (Intercept) -1.0771095 -0.81427 -0.62781 -0.51408 -0.23029
#> PEXPOSURE   -0.0004135  0.04738  0.07927  0.11564  0.20607
#> PCTAGE65P    2.5911599  3.20769  3.63005  4.09097  4.74100
#> PCTOWNHOME  -0.8037868 -0.54136 -0.35322 -0.20277  0.17058
#> 

# \dontrun{
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01177 0.01571 0.0002221      0.0008781
#> (Intercept) -0.79575 0.14274 0.0020186      0.0078442
#> PEXPOSURE    0.08036 0.02929 0.0004143      0.0017545
#> PCTAGE65P    3.81500 0.57129 0.0080793      0.0330363
#> PCTOWNHOME  -0.38426 0.15508 0.0021932      0.0087813
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.01899  0.00175  0.01220  0.02213  0.04328
#> (Intercept) -1.07330 -0.89063 -0.79366 -0.70020 -0.51399
#> PEXPOSURE    0.02200  0.06082  0.08123  0.10067  0.13675
#> PCTAGE65P    2.70738  3.41885  3.80976  4.20150  4.94589
#> PCTOWNHOME  -0.70906 -0.47931 -0.38839 -0.28703 -0.06345
#> 

esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)

#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 

res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02676 0.01581 0.0002236       0.001178
#> (Intercept)     -1.09750 0.25378 0.0035889       0.020707
#> PEXPOSURE        0.21726 0.08428 0.0011920       0.006773
#> PCTAGE65P        3.41184 0.64490 0.0091202       0.049637
#> PCTOWNHOME      -0.01373 0.22966 0.0032479       0.016678
#> lag.(Intercept)  0.12153 0.06099 0.0008625       0.005121
#> lag.PEXPOSURE   -0.03594 0.01713 0.0002422       0.001467
#> lag.PCTAGE65P    0.24843 0.21799 0.0030829       0.016826
#> lag.PCTOWNHOME  -0.14533 0.06036 0.0008536       0.004839
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> lambda          -0.004004  0.01652  0.02646  0.03706  0.058414
#> (Intercept)     -1.596596 -1.26665 -1.09711 -0.92951 -0.569075
#> PEXPOSURE        0.060047  0.15783  0.21404  0.27377  0.378642
#> PCTAGE65P        2.125474  2.99396  3.38901  3.80196  4.759975
#> PCTOWNHOME      -0.432544 -0.18179 -0.02072  0.14040  0.469260
#> lag.(Intercept)  0.003745  0.08293  0.12172  0.16425  0.246727
#> lag.PEXPOSURE   -0.069168 -0.04666 -0.03371 -0.02598 -0.003682
#> lag.PCTAGE65P   -0.204498  0.09091  0.26104  0.40582  0.663236
#> lag.PCTOWNHOME  -0.270422 -0.18572 -0.14366 -0.10596 -0.030349
#> 

lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 

res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03915 0.01459 0.0002064      0.0008488
#> (Intercept) -0.50125 0.15408 0.0021790      0.0088983
#> PEXPOSURE    0.04852 0.03298 0.0004664      0.0018774
#> PCTAGE65P    3.61157 0.63269 0.0089476      0.0392880
#> PCTOWNHOME  -0.42602 0.17245 0.0024389      0.0101946
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> rho          0.009349  0.02961  0.03954  0.04917  0.06723
#> (Intercept) -0.802174 -0.60725 -0.50419 -0.40635 -0.20369
#> PEXPOSURE   -0.015802  0.02608  0.04816  0.07157  0.10954
#> PCTAGE65P    2.432746  3.19362  3.62072  4.02246  4.87609
#> PCTOWNHOME  -0.768748 -0.53969 -0.42812 -0.31526 -0.06927
#> 

lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)

#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 

res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02682 0.01653 0.0002338       0.001224
#> (Intercept)     -1.12343 0.25876 0.0036595       0.020136
#> PEXPOSURE        0.21335 0.08094 0.0011446       0.006299
#> PCTAGE65P        3.36142 0.68400 0.0096733       0.052746
#> PCTOWNHOME       0.09177 0.26971 0.0038143       0.021966
#> lag.(Intercept)  0.14090 0.06108 0.0008638       0.005038
#> lag.PEXPOSURE   -0.03547 0.01604 0.0002268       0.001230
#> lag.PCTAGE65P    0.13664 0.23180 0.0032781       0.018428
#> lag.PCTOWNHOME  -0.15042 0.05940 0.0008400       0.004528
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> rho             -0.005181  0.01542  0.02730  0.03756  0.059973
#> (Intercept)     -1.618594 -1.29277 -1.12902 -0.95147 -0.583037
#> PEXPOSURE        0.060313  0.15941  0.21675  0.26835  0.378858
#> PCTAGE65P        2.091132  2.87639  3.37081  3.84396  4.618430
#> PCTOWNHOME      -0.439832 -0.09099  0.09702  0.28782  0.615503
#> lag.(Intercept)  0.016612  0.10280  0.14452  0.18005  0.256387
#> lag.PEXPOSURE   -0.066728 -0.04674 -0.03593 -0.02417 -0.006011
#> lag.PCTAGE65P   -0.306106 -0.02674  0.12948  0.29697  0.606519
#> lag.PCTOWNHOME  -0.270533 -0.18772 -0.14938 -0.11312 -0.041099
#> 

ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho         -0.09328 0.02549 0.0003605       0.002807
#> lambda       0.11428 0.01584 0.0002240       0.001540
#> (Intercept) -0.94694 0.23887 0.0033782       0.030263
#> PEXPOSURE    0.16912 0.06891 0.0009745       0.011881
#> PCTAGE65P    3.22440 0.63892 0.0090357       0.053036
#> PCTOWNHOME  -0.15911 0.21919 0.0030998       0.022289
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%     25%      50%      75%    97.5%
#> rho         -0.13825 -0.1107 -0.09389 -0.07871 -0.03551
#> lambda       0.07682  0.1072  0.11634  0.12503  0.13925
#> (Intercept) -1.39958 -1.1172 -0.93272 -0.77853 -0.51855
#> PEXPOSURE    0.03267  0.1221  0.17214  0.21368  0.30463
#> PCTAGE65P    2.07593  2.7990  3.21875  3.63317  4.55371
#> PCTOWNHOME  -0.57334 -0.3151 -0.15804 -0.01052  0.28834
#> 

ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)

#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 

res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)

#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                      Mean      SD  Naive SE Time-series SE
#> rho              0.007638 0.06310 0.0008924       0.012083
#> lambda           0.010881 0.06297 0.0008905       0.011265
#> (Intercept)     -1.104056 0.24469 0.0034604       0.017333
#> PEXPOSURE        0.215041 0.07676 0.0010855       0.005712
#> PCTAGE65P        3.368310 0.66683 0.0094303       0.049312
#> PCTOWNHOME       0.063564 0.25724 0.0036379       0.018881
#> lag.(Intercept)  0.117444 0.05966 0.0008437       0.004982
#> lag.PEXPOSURE   -0.033177 0.01658 0.0002345       0.001498
#> lag.PCTAGE65P    0.204594 0.28989 0.0040997       0.028322
#> lag.PCTOWNHOME  -0.142960 0.06371 0.0009011       0.005254
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%       25%      50%      75%     97.5%
#> rho             -0.114019 -0.045907  0.01171  0.06298  0.104308
#> lambda          -0.103577 -0.041704  0.01586  0.06374  0.111342
#> (Intercept)     -1.591350 -1.261629 -1.11569 -0.93784 -0.600490
#> PEXPOSURE        0.070314  0.162090  0.21453  0.26415  0.368288
#> PCTAGE65P        2.077900  2.926809  3.39183  3.81345  4.637470
#> PCTOWNHOME      -0.396024 -0.121030  0.05620  0.24993  0.554589
#> lag.(Intercept) -0.005675  0.082040  0.11754  0.15804  0.234216
#> lag.PEXPOSURE   -0.064986 -0.045013 -0.03387 -0.02116 -0.003128
#> lag.PCTAGE65P   -0.325625 -0.004242  0.19177  0.39492  0.755637
#> lag.PCTOWNHOME  -0.268095 -0.182600 -0.14094 -0.09725 -0.026733
#> 

# }