MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for Spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for Sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object

A spatial regression model object fitted by maximum likelihood with spautolm

mcmc

The number of MCMC iterations after burnin

verbose

default NULL, use global option value; if TRUE, reports progress

...

Arguments passed through

burnin

The number of burn-in iterations for the sampler

scale

a positive scale parameter

listw, listw2

listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control

list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

Author

Roger Bivand Roger.Bivand@nhh.no

See also

Examples

require("sf", quietly=TRUE)
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
#> [1] TRUE
#require("spdep", quietly=TRUE)
nyadjlw <- spdep::mat2listw(nyadjmat, as.character(nydata$AREAKEY))
listw_NY <- spdep::nb2listw(nyadjlw$neighbours, style="B")
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017209 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 
res <- MCMCsamp(esar1f, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04001 0.01707 0.0005399       0.001952
#> (Intercept) -0.64306 0.18227 0.0057638       0.021874
#> PEXPOSURE    0.07714 0.04402 0.0013921       0.005096
#> PCTAGE65P    3.77115 0.66208 0.0209367       0.078911
#> PCTOWNHOME  -0.41005 0.20560 0.0065016       0.025344
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.003697  0.02876  0.04096  0.05108  0.07377
#> (Intercept) -1.010162 -0.75178 -0.62651 -0.50945 -0.33213
#> PEXPOSURE   -0.012870  0.04708  0.08023  0.10886  0.15518
#> PCTAGE65P    2.377758  3.33159  3.82741  4.16499  5.08140
#> PCTOWNHOME  -0.763610 -0.56285 -0.41434 -0.28717  0.04188
#> 
# \dontrun{
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01432 0.01645 0.0002327      0.0009528
#> (Intercept) -0.82781 0.15037 0.0021266      0.0091370
#> PEXPOSURE    0.08532 0.03057 0.0004323      0.0018015
#> PCTAGE65P    3.82149 0.58470 0.0082689      0.0340386
#> PCTOWNHOME  -0.35000 0.16243 0.0022972      0.0091124
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%     75%    97.5%
#> lambda      -0.01695  0.003279  0.01484  0.0253  0.04706
#> (Intercept) -1.13038 -0.918410 -0.81946 -0.7302 -0.53934
#> PEXPOSURE    0.02943  0.064686  0.08326  0.1050  0.14704
#> PCTAGE65P    2.61095  3.438180  3.84949  4.1896  4.93237
#> PCTOWNHOME  -0.67443 -0.464258 -0.34844 -0.2349 -0.03659
#> 
ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 
res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08622 0.02920 0.0004129       0.001653
#> (Intercept) -0.68322 0.19789 0.0027987       0.011727
#> PEXPOSURE    0.08282 0.05247 0.0007421       0.003270
#> PCTAGE65P    3.75704 0.63493 0.0089793       0.036382
#> PCTOWNHOME  -0.35288 0.22991 0.0032515       0.014529
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%     25%      50%     75%    97.5%
#> lambda       0.02522  0.0658  0.08841  0.1086  0.13455
#> (Intercept) -1.08604 -0.8154 -0.67904 -0.5458 -0.30916
#> PEXPOSURE   -0.02079  0.0487  0.07909  0.1156  0.19020
#> PCTAGE65P    2.57120  3.2835  3.75527  4.2142  5.01250
#> PCTOWNHOME  -0.82748 -0.4938 -0.34169 -0.2025  0.09923
#> 
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01510 0.01693 0.0002394       0.001083
#> (Intercept) -0.81406 0.15443 0.0021840       0.009162
#> PEXPOSURE    0.08481 0.03188 0.0004509       0.001959
#> PCTAGE65P    3.81810 0.55635 0.0078680       0.030659
#> PCTOWNHOME  -0.37343 0.16429 0.0023234       0.010419
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%     50%      75%    97.5%
#> lambda      -0.01840  0.004274  0.0153  0.02593  0.04866
#> (Intercept) -1.13018 -0.923852 -0.8057 -0.70461 -0.53477
#> PEXPOSURE    0.02828  0.061177  0.0845  0.10433  0.15129
#> PCTAGE65P    2.71673  3.436993  3.8196  4.18692  4.87663
#> PCTOWNHOME  -0.67435 -0.492926 -0.3811 -0.26100 -0.03220
#> 
ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 
res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04093 0.04223 0.0005973       0.002928
#> (Intercept) -0.84314 0.19274 0.0027257       0.014578
#> PEXPOSURE    0.09392 0.04040 0.0005713       0.003673
#> PCTAGE65P    3.73730 0.58584 0.0082850       0.034082
#> PCTOWNHOME  -0.35679 0.18917 0.0026752       0.012137
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.04111  0.01262  0.04172  0.06947  0.12566
#> (Intercept) -1.28030 -0.94566 -0.82057 -0.71741 -0.52485
#> PEXPOSURE    0.02852  0.06854  0.08920  0.11372  0.19814
#> PCTAGE65P    2.60227  3.33711  3.73379  4.11492  4.95998
#> PCTOWNHOME  -0.69743 -0.48868 -0.36696 -0.23354  0.05357
#> 
# }
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04646 0.01589 0.0005025       0.002026
#> (Intercept) -0.62387 0.19521 0.0061731       0.028032
#> PEXPOSURE    0.07337 0.03977 0.0012575       0.004163
#> PCTAGE65P    3.77748 0.67223 0.0212579       0.078551
#> PCTOWNHOME  -0.42259 0.19976 0.0063168       0.027140
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.015933  0.03450  0.04683  0.05826  0.07461
#> (Intercept) -1.036934 -0.74613 -0.62918 -0.50290 -0.22351
#> PEXPOSURE    0.006434  0.04402  0.07379  0.09781  0.14820
#> PCTAGE65P    2.287060  3.30526  3.82691  4.21965  4.95288
#> PCTOWNHOME  -0.794631 -0.55436 -0.42774 -0.29149 -0.02868
#> 
# \dontrun{
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01264 0.01693 0.0002394      0.0009194
#> (Intercept) -0.79572 0.14919 0.0021098      0.0085953
#> PEXPOSURE    0.08218 0.02967 0.0004196      0.0017624
#> PCTAGE65P    3.77097 0.56713 0.0080204      0.0321372
#> PCTOWNHOME  -0.37425 0.15969 0.0022584      0.0095249
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.01898  0.00148  0.01220  0.02415  0.04619
#> (Intercept) -1.07804 -0.89673 -0.80093 -0.68897 -0.50028
#> PEXPOSURE    0.02375  0.06145  0.08308  0.10270  0.14032
#> PCTAGE65P    2.61747  3.40744  3.76515  4.13820  4.86343
#> PCTOWNHOME  -0.68124 -0.48248 -0.36610 -0.26440 -0.06796
#> 
esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 
res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02664 0.01731 0.0002448       0.001445
#> (Intercept)     -1.11779 0.24615 0.0034810       0.018477
#> PEXPOSURE        0.21666 0.07903 0.0011177       0.006156
#> PCTAGE65P        3.44403 0.66613 0.0094205       0.054200
#> PCTOWNHOME       0.02791 0.24636 0.0034841       0.017966
#> lag.(Intercept)  0.11899 0.05413 0.0007655       0.003895
#> lag.PEXPOSURE   -0.03408 0.01569 0.0002219       0.001188
#> lag.PCTAGE65P    0.26541 0.22248 0.0031464       0.017271
#> lag.PCTOWNHOME  -0.15371 0.05829 0.0008243       0.004495
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> lambda          -0.007983  0.01561  0.02677  0.03815  0.060478
#> (Intercept)     -1.630741 -1.26755 -1.12615 -0.95492 -0.658527
#> PEXPOSURE        0.060952  0.16303  0.21795  0.27110  0.369123
#> PCTAGE65P        2.179985  2.98375  3.46602  3.94588  4.721013
#> PCTOWNHOME      -0.454307 -0.13862  0.02277  0.17105  0.556899
#> lag.(Intercept)  0.009429  0.08303  0.11691  0.15236  0.225551
#> lag.PEXPOSURE   -0.064503 -0.04472 -0.03393 -0.02320 -0.006341
#> lag.PCTAGE65P   -0.207992  0.12362  0.25575  0.43532  0.690657
#> lag.PCTOWNHOME  -0.276539 -0.19200 -0.15699 -0.11334 -0.041910
#> 
lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 
res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03941 0.01438 0.0002033      0.0008153
#> (Intercept) -0.50459 0.16645 0.0023540      0.0103933
#> PEXPOSURE    0.04431 0.03807 0.0005383      0.0025311
#> PCTAGE65P    3.58732 0.63463 0.0089750      0.0395655
#> PCTOWNHOME  -0.40438 0.17782 0.0025147      0.0104785
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> rho          0.01139  0.02985  0.03940  0.04977  0.06749
#> (Intercept) -0.83475 -0.61489 -0.50587 -0.39908 -0.16192
#> PEXPOSURE   -0.03039  0.01908  0.04369  0.06867  0.12181
#> PCTAGE65P    2.40599  3.17100  3.60271  4.00568  4.84911
#> PCTOWNHOME  -0.75254 -0.51872 -0.40299 -0.28250 -0.06687
#> 
lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 
res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02639 0.01731 0.0002447       0.001376
#> (Intercept)     -1.09340 0.24932 0.0035259       0.018650
#> PEXPOSURE        0.21187 0.08271 0.0011697       0.006697
#> PCTAGE65P        3.33757 0.62894 0.0088945       0.046186
#> PCTOWNHOME       0.02806 0.25799 0.0036485       0.019903
#> lag.(Intercept)  0.12638 0.05644 0.0007982       0.004347
#> lag.PEXPOSURE   -0.03369 0.01644 0.0002324       0.001375
#> lag.PCTAGE65P    0.16969 0.23805 0.0033665       0.018638
#> lag.PCTOWNHOME  -0.13408 0.06147 0.0008693       0.004898
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%       25%      50%      75%     97.5%
#> rho             -0.009348  0.014863  0.02659  0.03817  0.059014
#> (Intercept)     -1.566398 -1.247599 -1.11066 -0.93527 -0.592916
#> PEXPOSURE        0.051143  0.154615  0.21446  0.26488  0.379335
#> PCTAGE65P        2.070955  2.946428  3.33766  3.75663  4.666186
#> PCTOWNHOME      -0.452715 -0.146853  0.02290  0.20161  0.563357
#> lag.(Intercept)  0.018492  0.088830  0.12357  0.16471  0.232746
#> lag.PEXPOSURE   -0.066219 -0.044733 -0.03338 -0.02204 -0.003191
#> lag.PCTAGE65P   -0.341037  0.001146  0.17672  0.35102  0.622026
#> lag.PCTOWNHOME  -0.254849 -0.177764 -0.13151 -0.08932 -0.020180
#> 
ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                  Mean      SD Naive SE Time-series SE
#> rho         -0.007526 0.08591 0.001215        0.02618
#> lambda       0.028126 0.08464 0.001197        0.02421
#> (Intercept) -0.632441 0.27560 0.003898        0.04505
#> PEXPOSURE    0.089391 0.07677 0.001086        0.01375
#> PCTAGE65P    3.296382 0.68594 0.009701        0.05142
#> PCTOWNHOME  -0.291066 0.21055 0.002978        0.02161
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%       50%      75%   97.5%
#> rho         -0.15552 -0.09457  0.008668  0.07076  0.1114
#> lambda      -0.12817 -0.04749  0.035462  0.11419  0.1398
#> (Intercept) -1.21573 -0.81705 -0.585620 -0.41254 -0.2203
#> PEXPOSURE   -0.01643  0.02928  0.070437  0.13724  0.2618
#> PCTAGE65P    2.06841  2.82620  3.299895  3.72438  4.6615
#> PCTOWNHOME  -0.67524 -0.43644 -0.296051 -0.15542  0.1733
#> 
ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                      Mean      SD  Naive SE Time-series SE
#> rho             -0.006243 0.07349 0.0010394       0.016978
#> lambda           0.023036 0.07006 0.0009907       0.014064
#> (Intercept)     -1.074786 0.25700 0.0036346       0.020044
#> PEXPOSURE        0.213653 0.07987 0.0011296       0.006275
#> PCTAGE65P        3.369138 0.64267 0.0090887       0.048578
#> PCTOWNHOME       0.035386 0.26783 0.0037877       0.021115
#> lag.(Intercept)  0.106733 0.06761 0.0009562       0.006909
#> lag.PEXPOSURE   -0.033731 0.01826 0.0002582       0.001782
#> lag.PCTAGE65P    0.239723 0.32512 0.0045978       0.037622
#> lag.PCTOWNHOME  -0.138544 0.07558 0.0010688       0.007509
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%       25%       50%      75%     97.5%
#> rho             -0.15259 -0.058974 -0.001855  0.05403  0.110005
#> lambda          -0.12122 -0.034805  0.030348  0.07995  0.127609
#> (Intercept)     -1.63521 -1.245895 -1.066876 -0.92132 -0.594111
#> PEXPOSURE        0.05949  0.156813  0.216836  0.26607  0.373617
#> PCTAGE65P        2.10893  2.932956  3.351803  3.78560  4.733524
#> PCTOWNHOME      -0.48089 -0.148442  0.034779  0.22580  0.583465
#> lag.(Intercept) -0.03405  0.065889  0.107361  0.15007  0.231268
#> lag.PEXPOSURE   -0.07338 -0.045326 -0.034166 -0.02251  0.004342
#> lag.PCTAGE65P   -0.39085 -0.002034  0.212151  0.47009  0.871781
#> lag.PCTOWNHOME  -0.29853 -0.183541 -0.139221 -0.08629  0.008422
#> 
# }