MCMC sample from fitted spatial regression

The MCMCsamp method uses rwmetrop, a random walk Metropolis algorithm, from LearnBayes to make MCMC samples from fitted maximum likelihood spatial regression models.

MCMCsamp(object, mcmc = 1L, verbose = NULL, ...)
# S3 method for Spautolm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
 burnin = 0L, scale=1, listw, control = list())
# S3 method for Sarlm
MCMCsamp(object, mcmc = 1L, verbose = NULL, ...,
    burnin=0L, scale=1, listw, listw2=NULL, control=list())

Arguments

object

A spatial regression model object fitted by maximum likelihood with spautolm

mcmc

The number of MCMC iterations after burnin

verbose

default NULL, use global option value; if TRUE, reports progress

...

Arguments passed through

burnin

The number of burn-in iterations for the sampler

scale

a positive scale parameter

listw, listw2

listw objects created for example by nb2listw; should be the same object(s) used for fitting the model

control

list of extra control arguments - see spautolm

Value

An object of class “mcmc” suited to coda, with attributes: “accept” acceptance rate; “type” input ML fitted model type “SAR”, “CAR”, “SMA”, “lag”, “mixed”, “error”, “sac”, “sacmixed”; “timings” run times

Note

If the acceptance rate is below 0.05, a warning will be issued; consider increasing mcmc.

References

Jim Albert (2007) Bayesian Computation with R, Springer, New York, pp. 104-105.

Author

Roger Bivand Roger.Bivand@nhh.no

See also

Examples

require("sf", quietly=TRUE)
nydata <- st_read(system.file("shapes/NY8_bna_utm18.gpkg", package="spData")[1], quiet=TRUE)
suppressMessages(nyadjmat <- as.matrix(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1])[-1]))
suppressMessages(ID <- as.character(names(foreign::read.dbf(system.file(
 "misc/nyadjwts.dbf", package="spData")[1]))[-1]))
identical(substring(ID, 2, 10), substring(as.character(nydata$AREAKEY), 2, 10))
#> [1] TRUE
#require("spdep", quietly=TRUE)
nyadjlw <- spdep::mat2listw(nyadjmat, as.character(nydata$AREAKEY))
listw_NY <- spdep::nb2listw(nyadjlw$neighbours, style="B")
esar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="SAR", method="eigen")
summary(esar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487 LR test value: 5.2438 p-value: 0.022026 
#> Numerical Hessian standard error of lambda: 0.017209 
#> 
#> Log likelihood: -276.1069 
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21
#> 
res <- MCMCsamp(esar1f, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04574 0.01631 0.0005157       0.001943
#> (Intercept) -0.64762 0.20439 0.0064632       0.028394
#> PEXPOSURE    0.07174 0.05229 0.0016535       0.007932
#> PCTAGE65P    3.79764 0.71373 0.0225701       0.097052
#> PCTOWNHOME  -0.37435 0.21462 0.0067870       0.028949
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%    97.5%
#> lambda       0.009512  0.03496  0.04562  0.05657  0.07601
#> (Intercept) -1.078533 -0.79115 -0.64525 -0.50379 -0.25764
#> PEXPOSURE   -0.011258  0.02982  0.07111  0.10114  0.17886
#> PCTAGE65P    2.296463  3.32782  3.79411  4.30275  5.23757
#> PCTOWNHOME  -0.816789 -0.52951 -0.37989 -0.22817  0.06022
#> 
# \dontrun{
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01268 0.01682 0.0002378      0.0009326
#> (Intercept) -0.80903 0.15630 0.0022104      0.0093344
#> PEXPOSURE    0.08053 0.03012 0.0004260      0.0016756
#> PCTAGE65P    3.85482 0.62946 0.0089019      0.0376144
#> PCTOWNHOME  -0.37460 0.16919 0.0023927      0.0094999
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.02048  0.001178  0.01244  0.02437  0.04530
#> (Intercept) -1.10731 -0.916103 -0.81074 -0.70054 -0.49788
#> PEXPOSURE    0.02297  0.060692  0.08015  0.09918  0.14477
#> PCTAGE65P    2.64213  3.426475  3.85592  4.30384  5.08299
#> PCTOWNHOME  -0.71918 -0.483336 -0.37333 -0.26035 -0.05803
#> 
ecar1f <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, family="CAR", method="eigen")
summary(ecar1f)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.539732 -0.384311 -0.030646  0.335126  3.808848 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.648362   0.181129 -3.5796 0.0003442
#> PEXPOSURE    0.077899   0.043692  1.7829 0.0745986
#> PCTAGE65P    3.703830   0.627185  5.9055 3.516e-09
#> PCTOWNHOME  -0.382789   0.195564 -1.9574 0.0503053
#> 
#> Lambda: 0.084123 LR test value: 5.8009 p-value: 0.016018 
#> Numerical Hessian standard error of lambda: 0.030868 
#> 
#> Log likelihood: -275.8283 
#> ML residual variance (sigma squared): 0.40758, (sigma: 0.63842)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 563.66
#> 
res <- MCMCsamp(ecar1f, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.08660 0.02799 0.0003958       0.001585
#> (Intercept) -0.66712 0.18998 0.0026867       0.010898
#> PEXPOSURE    0.08105 0.04945 0.0006993       0.002951
#> PCTAGE65P    3.69618 0.61706 0.0087266       0.034937
#> PCTOWNHOME  -0.36487 0.21258 0.0030063       0.012348
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%     75%   97.5%
#> lambda       0.024505  0.06984  0.08856  0.1053  0.1337
#> (Intercept) -1.077790 -0.78509 -0.65787 -0.5324 -0.3313
#> PEXPOSURE   -0.008748  0.04608  0.07881  0.1124  0.1848
#> PCTAGE65P    2.480087  3.29071  3.65030  4.0920  4.9640
#> PCTOWNHOME  -0.747823 -0.50889 -0.37474 -0.2253  0.0752
#> 
esar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="SAR", method="eigen")
summary(esar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "SAR", method = "eigen")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.48488 -0.26823  0.09489  0.46552  4.28343 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.797063   0.144054 -5.5331 3.146e-08
#> PEXPOSURE    0.080545   0.028334  2.8428  0.004473
#> PCTAGE65P    3.816731   0.576037  6.6258 3.453e-11
#> PCTOWNHOME  -0.380778   0.156507 -2.4330  0.014975
#> 
#> Lambda: 0.0095636 LR test value: 0.32665 p-value: 0.56764 
#> Numerical Hessian standard error of lambda: 0.016466 
#> 
#> Log likelihood: -251.6017 
#> ML residual variance (sigma squared): 1104.1, (sigma: 33.229)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.2
#> 
res <- MCMCsamp(esar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01260 0.01668 0.0002359      0.0009513
#> (Intercept) -0.80138 0.15486 0.0021901      0.0092477
#> PEXPOSURE    0.08342 0.03174 0.0004489      0.0019194
#> PCTAGE65P    3.79136 0.60836 0.0086035      0.0364965
#> PCTOWNHOME  -0.38036 0.17193 0.0024315      0.0106121
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%        25%      50%      75%    97.5%
#> lambda      -0.01943  0.0007966  0.01306  0.02364  0.04698
#> (Intercept) -1.11396 -0.9017662 -0.79900 -0.70075 -0.51453
#> PEXPOSURE    0.02137  0.0616812  0.08217  0.10428  0.14675
#> PCTAGE65P    2.51153  3.3955875  3.79538  4.19522  5.00877
#> PCTOWNHOME  -0.70265 -0.4932476 -0.38262 -0.26888 -0.03282
#> 
ecar1fw <- spautolm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8, family="CAR", method="eigen")
summary(ecar1fw)
#> 
#> Call: 
#> spautolm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, weights = POP8, family = "CAR", method = "eigen")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.491042 -0.270906  0.081435  0.451556  4.198134 
#> 
#> Coefficients: 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.790154   0.144862 -5.4545 4.910e-08
#> PEXPOSURE    0.081922   0.028593  2.8651  0.004169
#> PCTAGE65P    3.825858   0.577720  6.6223 3.536e-11
#> PCTOWNHOME  -0.386820   0.157436 -2.4570  0.014010
#> 
#> Lambda: 0.022419 LR test value: 0.38785 p-value: 0.53343 
#> Numerical Hessian standard error of lambda: 0.038543 
#> 
#> Log likelihood: -251.5711 
#> ML residual variance (sigma squared): 1102.9, (sigma: 33.21)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 515.14
#> 
res <- MCMCsamp(ecar1fw, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.03881 0.03972 0.0005618       0.002321
#> (Intercept) -0.82878 0.16388 0.0023176       0.009945
#> PEXPOSURE    0.08829 0.03218 0.0004551       0.001815
#> PCTAGE65P    3.69365 0.61853 0.0087473       0.038114
#> PCTOWNHOME  -0.34529 0.17423 0.0024640       0.009876
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%      25%      50%      75%    97.5%
#> lambda      -0.04871  0.01175  0.04257  0.06597  0.10687
#> (Intercept) -1.16203 -0.93101 -0.82343 -0.71461 -0.53938
#> PEXPOSURE    0.02561  0.06670  0.08821  0.10992  0.15509
#> PCTAGE65P    2.45544  3.31534  3.70597  4.11606  4.90850
#> PCTOWNHOME  -0.66483 -0.45960 -0.35619 -0.23668  0.02449
#> 
# }
esar0 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0, mcmc=1000, burnin=200, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:1000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.04359 0.01664 0.0005261       0.002186
#> (Intercept) -0.65653 0.19202 0.0060721       0.026630
#> PEXPOSURE    0.08753 0.04582 0.0014491       0.006245
#> PCTAGE65P    3.70175 0.63354 0.0200343       0.079600
#> PCTOWNHOME  -0.39934 0.19326 0.0061114       0.024507
#> 
#> 2. Quantiles for each variable:
#> 
#>                   2.5%      25%      50%      75%     97.5%
#> lambda       0.0070796  0.03240  0.04412  0.05555  0.072858
#> (Intercept) -1.0403101 -0.80153 -0.64237 -0.50713 -0.315161
#> PEXPOSURE    0.0003175  0.05607  0.08729  0.12122  0.178753
#> PCTAGE65P    2.3902483  3.31834  3.73186  4.07082  5.038157
#> PCTOWNHOME  -0.7520826 -0.54651 -0.37950 -0.26557 -0.009124
#> 
# \dontrun{
esar0w <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, weights=POP8)
summary(esar0)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.56754 -0.38239 -0.02643  0.33109  4.01219 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.618193   0.176784 -3.4969 0.0004707
#> PEXPOSURE    0.071014   0.042051  1.6888 0.0912635
#> PCTAGE65P    3.754200   0.624722  6.0094 1.862e-09
#> PCTOWNHOME  -0.419890   0.191329 -2.1946 0.0281930
#> 
#> Lambda: 0.040487, LR test value: 5.2438, p-value: 0.022026
#> Asymptotic standard error: 0.016214
#>     z-value: 2.4971, p-value: 0.01252
#> Wald statistic: 6.2356, p-value: 0.01252
#> 
#> Log likelihood: -276.1069 for error model
#> ML residual variance (sigma squared): 0.41388, (sigma: 0.64333)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 564.21, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(esar0w, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> lambda       0.01411 0.01633 0.0002309      0.0009941
#> (Intercept) -0.79110 0.15754 0.0022280      0.0094295
#> PEXPOSURE    0.08330 0.03121 0.0004414      0.0020430
#> PCTAGE65P    3.75371 0.61609 0.0087128      0.0434190
#> PCTOWNHOME  -0.38853 0.16547 0.0023401      0.0095368
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%       25%      50%      75%    97.5%
#> lambda      -0.01822  0.003018  0.01479  0.02476  0.04536
#> (Intercept) -1.09778 -0.897840 -0.79040 -0.67982 -0.48500
#> PEXPOSURE    0.02732  0.060445  0.08176  0.10426  0.14724
#> PCTAGE65P    2.43208  3.362637  3.78355  4.14790  4.98204
#> PCTOWNHOME  -0.69958 -0.505921 -0.39396 -0.28010 -0.04569
#> 
esar1 <- errorsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, etype="emixed")
summary(esar1)
#> 
#> Call:errorsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, 
#>     data = nydata, listw = listw_NY, etype = "emixed")
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -1.81562 -0.37641 -0.02224  0.33638  4.00054 
#> 
#> Type: error 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.118019   0.247425 -4.5186 6.225e-06
#> PEXPOSURE        0.218279   0.079245  2.7545  0.005879
#> PCTAGE65P        3.416477   0.645587  5.2920 1.210e-07
#> PCTOWNHOME       0.036593   0.249835  0.1465  0.883551
#> lag.(Intercept)  0.121515   0.057636  2.1083  0.035003
#> lag.PEXPOSURE   -0.035075   0.015943 -2.2000  0.027808
#> lag.PCTAGE65P    0.263096   0.220118  1.1953  0.231989
#> lag.PCTOWNHOME  -0.155680   0.059213 -2.6291  0.008560
#> 
#> Lambda: 0.022723, LR test value: 1.6846, p-value: 0.19432
#> Asymptotic standard error: 0.017169
#>     z-value: 1.3235, p-value: 0.18567
#> Wald statistic: 1.7516, p-value: 0.18567
#> 
#> Log likelihood: -269.5398 for error model
#> ML residual variance (sigma squared): 0.39759, (sigma: 0.63055)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 559.08, (AIC for lm: 558.76)
#> 
res <- MCMCsamp(esar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> lambda           0.02921 0.01632 0.0002308       0.001117
#> (Intercept)     -1.11531 0.25524 0.0036097       0.019366
#> PEXPOSURE        0.21952 0.08672 0.0012264       0.006817
#> PCTAGE65P        3.38613 0.65731 0.0092958       0.051229
#> PCTOWNHOME       0.02355 0.26782 0.0037876       0.022234
#> lag.(Intercept)  0.12018 0.06275 0.0008874       0.005015
#> lag.PEXPOSURE   -0.03573 0.01801 0.0002547       0.001490
#> lag.PCTAGE65P    0.25385 0.21938 0.0031025       0.016597
#> lag.PCTOWNHOME  -0.14803 0.06505 0.0009199       0.005328
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%     97.5%
#> lambda          -0.006801  0.01912  0.03039  0.04106  0.056907
#> (Intercept)     -1.620741 -1.27411 -1.11579 -0.95616 -0.629181
#> PEXPOSURE        0.058304  0.16164  0.21736  0.27726  0.404403
#> PCTAGE65P        2.067796  2.93733  3.39383  3.86946  4.642652
#> PCTOWNHOME      -0.485563 -0.15050  0.01703  0.18295  0.573443
#> lag.(Intercept) -0.012369  0.08033  0.11834  0.16449  0.242625
#> lag.PEXPOSURE   -0.071580 -0.04754 -0.03583 -0.02335 -0.001752
#> lag.PCTAGE65P   -0.160760  0.10978  0.24992  0.39586  0.685783
#> lag.PCTOWNHOME  -0.284280 -0.19053 -0.14957 -0.10386 -0.026001
#> 
lsar0 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(lsar0)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.586752 -0.391580 -0.022469  0.338017  4.029430 
#> 
#> Type: lag 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.514495   0.156154 -3.2948  0.000985
#> PEXPOSURE    0.047627   0.034509  1.3801  0.167542
#> PCTAGE65P    3.648198   0.599046  6.0900 1.129e-09
#> PCTOWNHOME  -0.414601   0.169554 -2.4453  0.014475
#> 
#> Rho: 0.038893, LR test value: 6.9683, p-value: 0.0082967
#> Asymptotic standard error: 0.015053
#>     z-value: 2.5837, p-value: 0.0097755
#> Wald statistic: 6.6754, p-value: 0.0097755
#> 
#> Log likelihood: -275.2447 for lag model
#> ML residual variance (sigma squared): 0.41166, (sigma: 0.6416)
#> Number of observations: 281 
#> Number of parameters estimated: 6 
#> AIC: 562.49, (AIC for lm: 567.46)
#> LM test for residual autocorrelation
#> test value: 1.4633, p-value: 0.22641
#> 
res <- MCMCsamp(lsar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho          0.03857 0.01468 0.0002076      0.0009627
#> (Intercept) -0.50537 0.15744 0.0022266      0.0095358
#> PEXPOSURE    0.05017 0.03495 0.0004943      0.0020160
#> PCTAGE65P    3.57242 0.64886 0.0091763      0.0401222
#> PCTOWNHOME  -0.41870 0.17201 0.0024325      0.0098024
#> 
#> 2. Quantiles for each variable:
#> 
#>                  2.5%      25%      50%      75%   97.5%
#> rho          0.007952  0.02964  0.03842  0.04884  0.0658
#> (Intercept) -0.811600 -0.60525 -0.50300 -0.41059 -0.1875
#> PEXPOSURE   -0.020593  0.02757  0.04976  0.07446  0.1172
#> PCTAGE65P    2.286965  3.14207  3.56910  4.01912  4.8538
#> PCTOWNHOME  -0.768824 -0.53004 -0.41432 -0.30221 -0.1012
#> 
lsar1 <- lagsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="mixed")
summary(lsar1)
#> 
#> Call:lagsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "mixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.799308 -0.390125 -0.021371  0.346128  3.965251 
#> 
#> Type: mixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.131233   0.249631 -4.5316 5.853e-06
#> PEXPOSURE        0.218364   0.079301  2.7536  0.005894
#> PCTAGE65P        3.361158   0.654123  5.1384 2.771e-07
#> PCTOWNHOME       0.071903   0.253967  0.2831  0.777085
#> lag.(Intercept)  0.132544   0.056175  2.3595  0.018300
#> lag.PEXPOSURE   -0.035239   0.015536 -2.2681  0.023322
#> lag.PCTAGE65P    0.161685   0.223690  0.7228  0.469798
#> lag.PCTOWNHOME  -0.140681   0.058529 -2.4036  0.016234
#> 
#> Rho: 0.026981, LR test value: 2.558, p-value: 0.10974
#> Asymptotic standard error: 0.016766
#>     z-value: 1.6093, p-value: 0.10755
#> Wald statistic: 2.5899, p-value: 0.10755
#> 
#> Log likelihood: -269.1031 for mixed model
#> ML residual variance (sigma squared): 0.39587, (sigma: 0.62918)
#> Number of observations: 281 
#> Number of parameters estimated: 10 
#> AIC: 558.21, (AIC for lm: 558.76)
#> LM test for residual autocorrelation
#> test value: 4.908, p-value: 0.026732
#> 
res <- MCMCsamp(lsar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho              0.02546 0.01560 0.0002206       0.001127
#> (Intercept)     -1.12291 0.23528 0.0033273       0.016750
#> PEXPOSURE        0.20664 0.08150 0.0011526       0.006440
#> PCTAGE65P        3.34306 0.63459 0.0089745       0.047635
#> PCTOWNHOME       0.08039 0.25824 0.0036521       0.019823
#> lag.(Intercept)  0.12992 0.05610 0.0007934       0.004184
#> lag.PEXPOSURE   -0.03237 0.01661 0.0002350       0.001349
#> lag.PCTAGE65P    0.17601 0.22020 0.0031141       0.017191
#> lag.PCTOWNHOME  -0.14485 0.06147 0.0008694       0.004898
#> 
#> 2. Quantiles for each variable:
#> 
#>                      2.5%      25%      50%      75%      97.5%
#> rho             -0.005633  0.01492  0.02583  0.03672  0.0546919
#> (Intercept)     -1.604527 -1.28298 -1.12321 -0.95678 -0.6829434
#> PEXPOSURE        0.050125  0.15027  0.20247  0.26256  0.3694802
#> PCTAGE65P        2.185329  2.87566  3.29640  3.77574  4.5956198
#> PCTOWNHOME      -0.405154 -0.10887  0.08084  0.26463  0.5621253
#> lag.(Intercept)  0.020801  0.09746  0.12877  0.16804  0.2423048
#> lag.PEXPOSURE   -0.064446 -0.04404 -0.03150 -0.02094 -0.0006196
#> lag.PCTAGE65P   -0.292292  0.02395  0.18511  0.33078  0.6106983
#> lag.PCTOWNHOME  -0.259346 -0.18777 -0.14775 -0.09857 -0.0345182
#> 
ssar0 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY)
summary(ssar0)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY)
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.468382 -0.375687 -0.034996  0.314714  3.833950 
#> 
#> Type: sac 
#> Coefficients: (asymptotic standard errors) 
#>              Estimate Std. Error z value  Pr(>|z|)
#> (Intercept) -0.386572   0.123188 -3.1381  0.001701
#> PEXPOSURE    0.026684   0.024013  1.1112  0.266479
#> PCTAGE65P    3.089824   0.562851  5.4896 4.029e-08
#> PCTOWNHOME  -0.323052   0.137449 -2.3503  0.018756
#> 
#> Rho: 0.089451
#> Asymptotic standard error: 0.019427
#>     z-value: 4.6046, p-value: 4.1325e-06
#> Lambda: -0.08192
#> Asymptotic standard error: 0.033201
#>     z-value: -2.4674, p-value: 0.01361
#> 
#> LR test value: 10.114, p-value: 0.0063661
#> 
#> Log likelihood: -273.672 for sac model
#> ML residual variance (sigma squared): 0.3766, (sigma: 0.61368)
#> Number of observations: 281 
#> Number of parameters estimated: 7 
#> AIC: 561.34, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar0, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD  Naive SE Time-series SE
#> rho         -0.08726 0.04442 0.0006282       0.008352
#> lambda       0.10819 0.03566 0.0005043       0.007967
#> (Intercept) -0.93807 0.26444 0.0037397       0.038693
#> PEXPOSURE    0.18248 0.08398 0.0011877       0.017843
#> PCTAGE65P    3.16666 0.63493 0.0089792       0.048137
#> PCTOWNHOME  -0.17718 0.25178 0.0035607       0.029326
#> 
#> 2. Quantiles for each variable:
#> 
#>                   2.5%     25%      50%      75%   97.5%
#> rho         -0.1599365 -0.1140 -0.09571 -0.07164  0.0419
#> lambda       0.0003195  0.1032  0.11857  0.12776  0.1410
#> (Intercept) -1.4219359 -1.1071 -0.94524 -0.76998 -0.3633
#> PEXPOSURE    0.0167477  0.1287  0.18413  0.24165  0.3376
#> PCTAGE65P    1.9113323  2.7419  3.16434  3.57044  4.5494
#> PCTOWNHOME  -0.6260033 -0.3754 -0.18575 -0.01325  0.3444
#> 
ssar1 <- sacsarlm(Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data=nydata,
 listw=listw_NY, type="sacmixed")
summary(ssar1)
#> 
#> Call:sacsarlm(formula = Z ~ PEXPOSURE + PCTAGE65P + PCTOWNHOME, data = nydata, 
#>     listw = listw_NY, type = "sacmixed")
#> 
#> Residuals:
#>       Min        1Q    Median        3Q       Max 
#> -1.633958 -0.363826 -0.019927  0.348238  3.655509 
#> 
#> Type: sacmixed 
#> Coefficients: (asymptotic standard errors) 
#>                  Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)     -1.133298   0.247495 -4.5791 4.670e-06
#> PEXPOSURE        0.206963   0.074480  2.7788  0.005456
#> PCTAGE65P        3.083983   0.671081  4.5955 4.316e-06
#> PCTOWNHOME       0.174800   0.256280  0.6821  0.495196
#> lag.(Intercept)  0.153427   0.050817  3.0192  0.002534
#> lag.PEXPOSURE   -0.033400   0.013817 -2.4173  0.015634
#> lag.PCTAGE65P   -0.079738   0.222144 -0.3589  0.719634
#> lag.PCTOWNHOME  -0.102502   0.056760 -1.8059  0.070940
#> 
#> Rho: 0.092495
#> Asymptotic standard error: 0.023829
#>     z-value: 3.8817, p-value: 0.00010375
#> Lambda: -0.091069
#> Asymptotic standard error: 0.038431
#>     z-value: -2.3697, p-value: 0.017804
#> 
#> LR test value: 22.379, p-value: 0.0010335
#> 
#> Log likelihood: -267.5392 for sacmixed model
#> ML residual variance (sigma squared): 0.35617, (sigma: 0.5968)
#> Number of observations: 281 
#> Number of parameters estimated: 11 
#> AIC: 557.08, (AIC for lm: 567.46)
#> 
res <- MCMCsamp(ssar1, mcmc=5000, burnin=500, listw=listw_NY)
summary(res)
#> 
#> Iterations = 1:5000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 5000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                     Mean      SD  Naive SE Time-series SE
#> rho             -0.02662 0.06591 0.0009321       0.017543
#> lambda           0.04322 0.06353 0.0008984       0.013112
#> (Intercept)     -1.04929 0.24066 0.0034034       0.018087
#> PEXPOSURE        0.20898 0.07496 0.0010601       0.005583
#> PCTAGE65P        3.33009 0.59137 0.0083632       0.042836
#> PCTOWNHOME      -0.02884 0.24555 0.0034726       0.017559
#> lag.(Intercept)  0.10258 0.06580 0.0009305       0.006671
#> lag.PEXPOSURE   -0.02897 0.01734 0.0002453       0.001601
#> lag.PCTAGE65P    0.24814 0.27466 0.0038843       0.030142
#> lag.PCTOWNHOME  -0.15597 0.06836 0.0009667       0.007458
#> 
#> 2. Quantiles for each variable:
#> 
#>                     2.5%       25%      50%      75%     97.5%
#> rho             -0.13031 -0.080683 -0.03651  0.02947  0.095047
#> lambda          -0.08952 -0.002219  0.06107  0.09443  0.126147
#> (Intercept)     -1.48935 -1.216458 -1.04203 -0.88493 -0.563809
#> PEXPOSURE        0.05571  0.158725  0.20941  0.26098  0.354936
#> PCTAGE65P        2.09479  2.964355  3.33650  3.70857  4.495449
#> PCTOWNHOME      -0.49141 -0.189202 -0.02056  0.14210  0.447350
#> lag.(Intercept) -0.02705  0.058357  0.10576  0.14826  0.229226
#> lag.PEXPOSURE   -0.06386 -0.040131 -0.03040 -0.01711  0.006348
#> lag.PCTAGE65P   -0.24420  0.054156  0.23806  0.45494  0.773841
#> lag.PCTOWNHOME  -0.28625 -0.203278 -0.15671 -0.10612 -0.030283
#> 
# }