Bayesian MCMC spatial simultaneous autoregressive model estimation

The spBreg_lag function is an early-release version of the Matlab Spatial Econometrics Toolbox function sar_g.m, using drawing by inversion, and not accommodating heteroskedastic disturbances.

spBreg_lag(formula, data = list(), listw, na.action, Durbin, type,
    zero.policy=NULL, control=list())
spBreg_sac(formula, data = list(), listw, listw2=NULL, na.action,
    Durbin, type, zero.policy=NULL, control=list())
spBreg_err(formula, data = list(), listw, na.action, Durbin, etype,
    zero.policy=NULL, control=list())
# S3 method for MCMC_sar_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for MCMC_sem_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)
# S3 method for MCMC_sac_g
impacts(obj, ..., tr=NULL, listw=NULL, evalues=NULL, Q=NULL)

Arguments

formula
a symbolic description of the model to be fit. The details of model specification are given for lm()

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which the function is called.

listw, listw2
a listw object created for example by nb2listw

na.action
a function (default options("na.action")), can also be na.omit or na.exclude with consequences for residuals and fitted values - in these cases the weights list will be subsetted to remove NAs in the data. It may be necessary to set zero.policy to TRUE because this subsetting may create no-neighbour observations. Note that only weights lists created without using the glist argument to nb2listw may be subsetted.

Durbin
default FALSE (spatial lag model); if TRUE, full spatial Durbin model; if a formula object, the subset of explanatory variables to lag

type, etype
(use the ‘Durbin=’ argument - retained for backwards compatibility only) default "lag", may be set to "mixed"; when "mixed", the lagged intercept is dropped for spatial weights style "W", that is row-standardised weights, but otherwise included; “Durbin” may be used instead of “mixed”

zero.policy
default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE (default) assign NA

control
list of extra control arguments - see section below

obj
A spatial regression object

…
Arguments passed through to methods in the coda package

tr
A vector of traces of powers of the spatial weights matrix created using trW, for approximate impact measures; if not given, listw must be given for exact measures (for small to moderate spatial weights matrices); the traces must be for the same spatial weights as were used in fitting the spatial regression, and must be row-standardised

evalues
vector of eigenvalues of spatial weights matrix for impacts calculations

Q
default NULL, else an integer number of cumulative power series impacts to calculate if tr is given

Control arguments

tol.opt:

the desired accuracy of the optimization - passed to optimize() (default=square root of double precision machine tolerance, a larger root may be used needed, see help(boston) for an example)

fdHess:

default NULL, then set to (method != "eigen") internally; use fdHess to compute an approximate Hessian using finite differences when using sparse matrix methods; used to make a coefficient covariance matrix when the number of observations is large; may be turned off to save resources if need be

optimHess:

default FALSE, use fdHess from nlme, if TRUE, use optim to calculate Hessian at optimum

optimHessMethod:

default “optimHess”, may be “nlm” or one of the optim methods

compiled_sse:

default FALSE; logical value used in the log likelihood function to choose compiled code for computing SSE

Imult:

default 2; used for preparing the Cholesky decompositions for updating in the Jacobian function

super:

if NULL (default), set to FALSE to use a simplicial decomposition for the sparse Cholesky decomposition and method “Matrix_J”, set to as.logical(NA) for method “Matrix”, if TRUE, use a supernodal decomposition

cheb_q:

default 5; highest power of the approximating polynomial for the Chebyshev approximation

MC_p:

default 16; number of random variates

MC_m:

default 30; number of products of random variates matrix and spatial weights matrix

spamPivot:

default “MMD”, alternative “RCM”

in_coef

default 0.1, coefficient value for initial Cholesky decomposition in “spam_update”

type

default “MC”, used with method “moments”; alternatives “mult” and “moments”, for use if trs is missing, trW

correct

default TRUE, used with method “moments” to compute the Smirnov/Anselin correction term

trunc

default TRUE, used with method “moments” to truncate the Smirnov/Anselin correction term

SE_method

default “LU”, may be “MC”

nrho

default 200, as in SE toolbox; the size of the first stage lndet grid; it may be reduced to for example 40

interpn

default 2000, as in SE toolbox; the size of the second stage lndet grid

small_asy

default TRUE; if the method is not “eigen”, use asymmetric covariances rather than numerical Hessian ones if n <= small

small

default 1500; threshold number of observations for asymmetric covariances when the method is not “eigen”

SElndet

default NULL, may be used to pass a pre-computed SE toolbox style matrix of coefficients and their lndet values to the "SE_classic" and "SE_whichMin" methods

LU_order

default FALSE; used in “LU_prepermutate”, note warnings given for lu method

pre_eig

default NULL; may be used to pass a pre-computed vector of eigenvalues

OrdVsign

default 1; used to set the sign of the final component to negative if -1 (alpha times ((sigma squared) squared) in Ord (1975) equation B.1).

Extra Bayesian control arguments

ldet_method

default “SE_classic”; equivalent to the method argument in lagsarlm

interval

default c(-1, 1); used unmodified or set internally by jacobianSetup

ndraw

default 2500L; integer total number of draws

nomit

default 500L; integer total number of omitted burn-in draws

thin

default 1L; integer thinning proportion

verbose

default FALSE; inverse of quiet argument in lagsarlm

detval

default NULL; not yet in use, precomputed matrix of log determinants

prior

a list with the following components:

rhoMH

default FALSE; use Metropolis or griddy Gibbs

Tbeta

default NULL; values of the betas variance-covariance matrix, set to diag(k)*1e+12 if NULL

c_beta

default NULL; values of the betas set to 0 if NULL

rho

default 0.5; value of the autoregressive coefficient

sige

default 1; value of the residual variance

nu

default 0; informative Gamma(nu,d0) prior on sige

d0

default 0; informative Gamma(nu,d0) prior on sige

a1

default 1.01; parameter for beta(a1,a2) prior on rho

a2

default 1.01; parameter for beta(a1,a2) prior on rho

cc

default 0.2; initial tuning parameter for M-H sampling

gG_sige

default TRUE; include sige in lambda griddy Gibbs update

cc1

default 0.2; initial tuning parameter for M-H sampling

cc2

default 0.2; initial tuning parameter for M-H sampling

References

LeSage J and RK Pace (2009) Introduction to Spatial Econometrics. CRC Press, Boca Raton.

Examples

#require("spdep", quietly=TRUE)
data(oldcol, package="spdep")
lw <- spdep::nb2listw(COL.nb, style="W")
ev <- eigenw(lw)
W <- as(lw, "CsparseMatrix")
trMatc <- trW(W, type="mult")
require("coda", quietly=TRUE)
set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw)
print(summary(COL.err.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  59.6663  6.99423 0.156396       0.156396
#> INC          -0.9421  0.39894 0.008921       0.008921
#> HOVAL        -0.3011  0.09984 0.002233       0.002233
#> lambda        0.5664  0.15362 0.003435       0.003435
#> sige        115.0482 25.96435 0.580581       0.622638
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%    97.5%
#> (Intercept) 45.6439 55.3852  60.1164  64.5665  72.1349
#> INC         -1.7360 -1.2047  -0.9445  -0.6803  -0.1379
#> HOVAL       -0.4977 -0.3683  -0.3010  -0.2346  -0.1070
#> lambda       0.2335  0.4717   0.5773   0.6758   0.8309
#> sige        74.4245 96.7013 111.9152 129.2463 176.8133
#> 

print(raftery.diag(COL.err.Bayes, r=0.01))

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 3        1052  937          1.120     
#>  INC         2        969   937          1.030     
#>  HOVAL       3        1052  937          1.120     
#>  lambda      3        1010  937          1.080     
#>  sige        2        930   937          0.993     
#> 

set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 control=list(prior=list(lambdaMH=TRUE)))
print(summary(COL.err.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> (Intercept)  59.5304  8.9916 0.201059       0.275885
#> INC          -0.9142  0.3969 0.008874       0.011266
#> HOVAL        -0.3027  0.1021 0.002282       0.002282
#> lambda        0.6011  0.1502 0.003358       0.007723
#> sige        116.6773 27.9707 0.625443       0.660902
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%    25%      50%      75%    97.5%
#> (Intercept) 44.0607 54.802  59.6796  64.2473  72.9189
#> INC         -1.6866 -1.184  -0.9177  -0.6422  -0.1429
#> HOVAL       -0.5102 -0.370  -0.3005  -0.2359  -0.1019
#> lambda       0.2826  0.511   0.6135   0.7023   0.8676
#> sige        74.1788 97.167 113.2955 130.2498 182.6024
#> 

print(raftery.diag(COL.err.Bayes, r=0.01))

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 3        1143  937          1.220     
#>  INC         2        969   937          1.030     
#>  HOVAL       2        892   937          0.952     
#>  lambda      17       4454  937          4.750     
#>  sige        2        930   937          0.993     
#> 

set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=TRUE)
print(summary(COL.err.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                  Mean      SD Naive SE Time-series SE
#> (Intercept)  72.66455 11.7701 0.263187       0.263187
#> INC          -1.01543  0.3763 0.008414       0.008414
#> HOVAL        -0.28060  0.1050 0.002347       0.002347
#> lag.INC      -1.06754  0.7388 0.016519       0.016519
#> lag.HOVAL     0.08961  0.2383 0.005328       0.005328
#> lambda        0.48878  0.1732 0.003873       0.003873
#> sige        114.66234 26.8527 0.600444       0.659903
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%      25%       50%      75%     97.5%
#> (Intercept) 48.2182 65.57218  73.00283  80.0588  95.67986
#> INC         -1.7264 -1.27249  -1.01848  -0.7658  -0.26325
#> HOVAL       -0.4835 -0.35298  -0.27998  -0.2105  -0.07685
#> lag.INC     -2.4553 -1.54934  -1.08873  -0.6092   0.39927
#> lag.HOVAL   -0.3926 -0.06243   0.09177   0.2443   0.55929
#> lambda       0.1165  0.37669   0.49775   0.6118   0.79690
#> sige        73.4827 95.70364 110.68368 129.1459 177.97468
#> 

print(summary(impacts(COL.err.Bayes)))

#> Impact measures (SDEM, MCMC, n):
#>           Direct    Indirect      Total
#> INC   -1.0154294 -1.06753527 -2.0829647
#> HOVAL -0.2806034  0.08960515 -0.1909982
#> ========================================================
#> Standard errors:
#>           Direct  Indirect     Total
#> INC   0.35656492 0.7000652 0.8257127
#> HOVAL 0.09946104 0.2258118 0.2716818
#> ========================================================
#> Z-values:
#>          Direct   Indirect      Total
#> INC   -2.847811 -1.5249083 -2.5226265
#> HOVAL -2.821239  0.3968134 -0.7030219
#> 
#> p-values:
#>       Direct    Indirect Total   
#> INC   0.0044021 0.12728  0.011648
#> HOVAL 0.0047839 0.69151  0.482042
#> 

print(raftery.diag(COL.err.Bayes, r=0.01))

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 2        930   937          0.993     
#>  INC         3        1010  937          1.080     
#>  HOVAL       2        930   937          0.993     
#>  lag.INC     2        892   937          0.952     
#>  lag.HOVAL   2        892   937          0.952     
#>  lambda      2        930   937          0.993     
#>  sige        2        930   937          0.993     
#> 

set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=TRUE, control=list(prior=list(lambdaMH=TRUE)))
print(summary(COL.err.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                  Mean      SD Naive SE Time-series SE
#> (Intercept)  70.40019 70.4345 1.574962       1.385430
#> INC          -0.97037  0.4110 0.009190       0.009823
#> HOVAL        -0.28455  0.1145 0.002561       0.002561
#> lag.INC      -0.95215  0.8426 0.018842       0.027001
#> lag.HOVAL     0.06555  0.2679 0.005990       0.007555
#> lambda        0.60702  0.1781 0.003983       0.014276
#> sige        120.45188 29.1908 0.652727       1.279486
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%       50%      75%     97.5%
#> (Intercept) 37.9440 62.5478  71.80756  80.9018 104.95272
#> INC         -1.7687 -1.2346  -0.98524  -0.7063  -0.14242
#> HOVAL       -0.5141 -0.3592  -0.28460  -0.2088  -0.06554
#> lag.INC     -2.6086 -1.5191  -0.95619  -0.4294   0.72268
#> lag.HOVAL   -0.4869 -0.1027   0.06799   0.2443   0.56913
#> lambda       0.2366  0.5043   0.62080   0.7333   0.92570
#> sige        76.7686 99.6498 116.80421 135.9061 189.12857
#> 

print(summary(impacts(COL.err.Bayes)))

#> Impact measures (SDEM, MCMC, n):
#>           Direct    Indirect      Total
#> INC   -0.9703690 -0.95214894 -1.9225179
#> HOVAL -0.2845537  0.06555386 -0.2189999
#> ========================================================
#> Standard errors:
#>          Direct  Indirect     Total
#> INC   0.3894378 0.7984944 0.9933065
#> HOVAL 0.1085247 0.2538476 0.3175629
#> ========================================================
#> Z-values:
#>          Direct  Indirect      Total
#> INC   -2.491717 -1.192430 -1.9354730
#> HOVAL -2.622018  0.258241 -0.6896269
#> 
#> p-values:
#>       Direct    Indirect Total   
#> INC   0.0127127 0.23309  0.052932
#> HOVAL 0.0087411 0.79622  0.490429
#> 

print(raftery.diag(COL.err.Bayes, r=0.01))

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 5        1472  937          1.570     
#>  INC         2        892   937          0.952     
#>  HOVAL       3        1010  937          1.080     
#>  lag.INC     3        1010  937          1.080     
#>  lag.HOVAL   2        930   937          0.993     
#>  lambda      14       3645  937          3.890     
#>  sige        2        930   937          0.993     
#> 

set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=~INC)
print(summary(COL.err.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  74.5423 10.98931 0.245728       0.245728
#> INC          -1.0339  0.38293 0.008563       0.008987
#> HOVAL        -0.2860  0.09857 0.002204       0.002204
#> lag.INC      -0.9190  0.59170 0.013231       0.012507
#> lambda        0.4879  0.17196 0.003845       0.003672
#> sige        112.0658 25.47022 0.569531       0.567228
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%     97.5%
#> (Intercept) 52.0905 67.8363  74.9216  81.7016  94.97558
#> INC         -1.7796 -1.2965  -1.0425  -0.7816  -0.23925
#> HOVAL       -0.4762 -0.3494  -0.2861  -0.2205  -0.08885
#> lag.INC     -2.0736 -1.3191  -0.9217  -0.5457   0.27871
#> lambda       0.1355  0.3737   0.4987   0.6088   0.80100
#> sige        73.0813 94.1656 108.2980 125.8857 173.13082
#> 

print(summary(impacts(COL.err.Bayes)))

#> Impact measures (SDEM, MCMC, n):
#>           Direct   Indirect      Total
#> INC   -1.0339443 -0.9190497 -1.9529940
#> HOVAL -0.2860029         NA -0.2860029
#> ========================================================
#> Standard errors:
#>           Direct  Indirect      Total
#> INC   0.36696535 0.5670367 0.69945670
#> HOVAL 0.09446444        NA 0.09446444
#> ========================================================
#> Z-values:
#>          Direct  Indirect     Total
#> INC   -2.817553 -1.620794 -2.792159
#> HOVAL -3.027625        NA -3.027625
#> 
#> p-values:
#>       Direct    Indirect Total    
#> INC   0.0048391 0.10506  0.0052358
#> HOVAL 0.0024648 NA       0.0024648
#> 

print(raftery.diag(COL.err.Bayes, r=0.01))

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 3        1010  937          1.080     
#>  INC         2        892   937          0.952     
#>  HOVAL       2        930   937          0.993     
#>  lag.INC     2        892   937          0.952     
#>  lambda      2        892   937          0.952     
#>  sige        3        1010  937          1.080     
#> 

set.seed(1)
COL.err.Bayes <- spBreg_err(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=~INC, control=list(prior=list(lambdaMH=TRUE)))
print(summary(COL.err.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean      SD Naive SE Time-series SE
#> (Intercept)  72.8864 13.2070 0.295318       0.332323
#> INC          -0.9843  0.3831 0.008566       0.009768
#> HOVAL        -0.2889  0.1008 0.002253       0.002253
#> lag.INC      -0.8606  0.6640 0.014847       0.014847
#> lambda        0.5779  0.1662 0.003715       0.010526
#> sige        114.9801 26.2240 0.586387       0.665850
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%    97.5%
#> (Intercept) 45.5280 65.3717  73.3929  81.2895  97.2255
#> INC         -1.7170 -1.2373  -0.9898  -0.7261  -0.2197
#> HOVAL       -0.4870 -0.3565  -0.2850  -0.2177  -0.1032
#> lag.INC     -2.1579 -1.2914  -0.8847  -0.4386   0.4898
#> lambda       0.2436  0.4725   0.5835   0.6981   0.8672
#> sige        73.5737 96.1297 111.5039 130.5606 176.0332
#> 

print(summary(impacts(COL.err.Bayes)))

#> Impact measures (SDEM, MCMC, n):
#>           Direct   Indirect      Total
#> INC   -0.9843411 -0.8605551 -1.8448962
#> HOVAL -0.2889068         NA -0.2889068
#> ========================================================
#> Standard errors:
#>           Direct  Indirect      Total
#> INC   0.36712699 0.6362974 0.80457985
#> HOVAL 0.09657666        NA 0.09657666
#> ========================================================
#> Z-values:
#>          Direct  Indirect     Total
#> INC   -2.681201 -1.352442 -2.292993
#> HOVAL -2.991476        NA -2.991476
#> 
#> p-values:
#>       Direct    Indirect Total    
#> INC   0.0073359 0.17623  0.0218484
#> HOVAL 0.0027763 NA       0.0027763
#> 

print(raftery.diag(COL.err.Bayes, r=0.01))

#> 
#> Quantile (q) = 0.025
#> Accuracy (r) = +/- 0.01
#> Probability (s) = 0.95 
#>                                                    
#>              Burn-in  Total Lower bound  Dependence
#>              (M)      (N)   (Nmin)       factor (I)
#>  (Intercept) 4        1192  937          1.270     
#>  INC         2        892   937          0.952     
#>  HOVAL       2        930   937          0.993     
#>  lag.INC     2        892   937          0.952     
#>  lambda      10       2853  937          3.040     
#>  sige        2        930   937          0.993     
#> 

set.seed(1)
COL.sacW.B0 <- spBreg_sac(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=FALSE, control=list(ndraw=1500L, nomit=500L))
print(summary(COL.sacW.B0))

#> 
#> Iterations = 501:1500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  49.3672  9.23725 0.292108       0.847100
#> INC          -0.9925  0.35777 0.011314       0.012624
#> HOVAL        -0.2847  0.09798 0.003099       0.003307
#> rho           0.3117  0.19633 0.006208       0.020392
#> lambda        0.1927  0.27130 0.008579       0.027162
#> sige        105.0462 23.96805 0.757936       0.757936
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%      25%      50%      75%    97.5%
#> (Intercept) 31.8783 43.12583  48.9398  55.3573  68.1858
#> INC         -1.6817 -1.24459  -0.9872  -0.7516  -0.3168
#> HOVAL       -0.4952 -0.34433  -0.2827  -0.2201  -0.1028
#> rho         -0.1548  0.21285   0.3406   0.4438   0.6176
#> lambda      -0.3788  0.01455   0.2171   0.3794   0.6988
#> sige        68.5025 88.13909 101.6888 118.2545 160.2792
#> 

print(summary(impacts(COL.sacW.B0, tr=trMatc), zstats=TRUE, short=TRUE))

#> Impact measures (sac, trace):
#>          Direct   Indirect      Total
#> INC   -1.017325 -0.4246675 -1.4419930
#> HOVAL -0.291816 -0.1218143 -0.4136303
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3668822 0.4180525 0.6533395
#> HOVAL 0.1016518 0.1341689 0.1983041
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -2.808257 -1.219897 -2.357550
#> HOVAL -2.910868 -1.106171 -2.240542
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0049810 0.22250  0.018396
#> HOVAL 0.0036043 0.26865  0.025056

set.seed(1)
COL.sacW.B1 <- spBreg_sac(CRIME ~ INC + HOVAL, data=COL.OLD, listw=lw,
 Durbin=TRUE, control=list(ndraw=1500L, nomit=500L))
print(summary(COL.sacW.B1))

#> 
#> Iterations = 501:1500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 1000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  60.1833 24.14827 0.763635       3.169760
#> INC          -0.9880  0.36251 0.011464       0.015819
#> HOVAL        -0.2839  0.09749 0.003083       0.003433
#> lag.INC      -0.8740  0.77272 0.024435       0.061565
#> lag.HOVAL     0.1704  0.22184 0.007015       0.013821
#> rho           0.1808  0.32007 0.010121       0.044327
#> lambda        0.2106  0.32071 0.010142       0.041657
#> sige        100.9463 24.32264 0.769149       0.946662
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%       25%     50%      75%    97.5%
#> (Intercept) 22.2042 42.277313 56.1007  76.4887 113.5316
#> INC         -1.7045 -1.223372 -0.9842  -0.7329  -0.3299
#> HOVAL       -0.4803 -0.349479 -0.2841  -0.2152  -0.1085
#> lag.INC     -2.5306 -1.336778 -0.8386  -0.3451   0.5156
#> lag.HOVAL   -0.3031  0.026900  0.1851   0.3249   0.5752
#> rho         -0.5076 -0.021227  0.2386   0.4320   0.6655
#> lambda      -0.5106 -0.007594  0.2129   0.4482   0.7389
#> sige        64.6867 84.395156 96.9599 114.4101 159.7252
#> 

print(summary(impacts(COL.sacW.B1, tr=trMatc), zstats=TRUE, short=TRUE))

#> Impact measures (sacmixed, trace):
#>           Direct  Indirect      Total
#> INC   -1.0332779 -1.239705 -2.2729830
#> HOVAL -0.2787547  0.140259 -0.1384957
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3510687 0.9383466 1.0337787
#> HOVAL 0.1002385 0.2776460 0.3184971
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.962199 -1.4351185 -2.3085930
#> HOVAL -2.775887  0.5415921 -0.4015108
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0030545 0.15125  0.020966
#> HOVAL 0.0055051 0.58810  0.688044

set.seed(1)
COL.lag.Bayes <- spBreg_lag(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw=lw)
print(summary(COL.lag.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  45.8256  8.04911 0.179984       0.179984
#> INC          -1.0459  0.34746 0.007770       0.007770
#> HOVAL        -0.2662  0.09366 0.002094       0.002094
#> rho           0.4146  0.12434 0.002780       0.002780
#> sige        107.8827 23.75095 0.531087       0.561152
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%     25%      50%      75%     97.5%
#> (Intercept) 30.3334 40.3140  45.9376  51.0903  61.51008
#> INC         -1.7252 -1.2774  -1.0477  -0.8158  -0.35423
#> HOVAL       -0.4533 -0.3312  -0.2686  -0.2019  -0.08107
#> rho          0.1645  0.3327   0.4157   0.5008   0.64982
#> sige        70.5778 90.6316 104.8556 120.4706 164.75421
#> 

print(summary(impacts(COL.lag.Bayes, tr=trMatc), short=TRUE, zstats=TRUE))

#> Impact measures (lag, trace):
#>           Direct   Indirect      Total
#> INC   -1.0962113 -0.6903627 -1.7865741
#> HOVAL -0.2790451 -0.1757347 -0.4547797
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3680189 0.4927553 0.7674941
#> HOVAL 0.1002735 0.1393959 0.2171077
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -3.003178 -1.554252 -2.437925
#> HOVAL -2.809370 -1.425055 -2.212507
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0026718 0.12012  0.014772
#> HOVAL 0.0049639 0.15414  0.026932

print(summary(impacts(COL.lag.Bayes, evalues=ev), short=TRUE, zstats=TRUE))

#> Impact measures (lag, evalues):
#>           Direct   Indirect      Total
#> INC   -1.0962113 -0.6903627 -1.7865741
#> HOVAL -0.2790451 -0.1757347 -0.4547797
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>          Direct  Indirect     Total
#> INC   0.3680193 0.4929524 0.7676307
#> HOVAL 0.1002746 0.1395496 0.2172252
#> 
#> Simulated z-values:
#>          Direct  Indirect     Total
#> INC   -3.003178 -1.553686 -2.437527
#> HOVAL -2.809345 -1.423571 -2.211367
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0026718 0.12026  0.014788
#> HOVAL 0.0049642 0.15457  0.027010

set.seed(1)
COL.D0.Bayes <- spBreg_lag(CRIME ~ INC + HOVAL, data=COL.OLD,
 listw=lw, Durbin=TRUE)
print(summary(COL.D0.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  45.3106 13.74660 0.307383       0.307383
#> INC          -0.9255  0.36779 0.008224       0.008224
#> HOVAL        -0.2941  0.09628 0.002153       0.002153
#> lag.INC      -0.5834  0.63860 0.014279       0.014279
#> lag.HOVAL     0.2396  0.19206 0.004295       0.004295
#> rho           0.3932  0.16296 0.003644       0.003644
#> sige        109.7428 24.97328 0.558419       0.607251
#> 
#> 2. Quantiles for each variable:
#> 
#>                 2.5%     25%      50%      75%    97.5%
#> (Intercept) 19.18523 35.9952  45.0347  54.5253  72.8545
#> INC         -1.63995 -1.1761  -0.9206  -0.6651  -0.2148
#> HOVAL       -0.48230 -0.3566  -0.2931  -0.2308  -0.1021
#> lag.INC     -1.82283 -1.0066  -0.5712  -0.1763   0.6692
#> lag.HOVAL   -0.14391  0.1068   0.2393   0.3670   0.6168
#> rho          0.06053  0.2826   0.3977   0.5078   0.6949
#> sige        71.19421 91.4872 106.1690 123.7826 170.9503
#> 

print(summary(impacts(COL.D0.Bayes, tr=trMatc), short=TRUE, zstats=TRUE))

#> Impact measures (mixed, trace):
#>          Direct   Indirect       Total
#> INC   -1.027773 -1.4589793 -2.48675192
#> HOVAL -0.280732  0.1908138 -0.08991828
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>           Direct  Indirect     Total
#> INC   0.38264314 1.3114491 1.4534522
#> HOVAL 0.09937135 0.3347428 0.3698437
#> 
#> Simulated z-values:
#>          Direct   Indirect      Total
#> INC   -2.739321 -1.2582210 -1.8564594
#> HOVAL -2.838369  0.5610536 -0.2548209
#> 
#> Simulated p-values:
#>       Direct    Indirect Total   
#> INC   0.0061566 0.20831  0.063388
#> HOVAL 0.0045345 0.57476  0.798861

set.seed(1)
COL.D1.Bayes <- spBreg_lag(CRIME ~ DISCBD + INC + HOVAL, data=COL.OLD,
 listw=lw, Durbin= ~ INC)
print(summary(COL.D1.Bayes))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                 Mean       SD Naive SE Time-series SE
#> (Intercept)  56.4447 13.43635 0.300446       0.300446
#> DISCBD       -4.7748  2.14395 0.047940       0.047940
#> INC          -0.9493  0.34681 0.007755       0.007755
#> HOVAL        -0.1798  0.09983 0.002232       0.002232
#> lag.INC       0.4262  0.63352 0.014166       0.014166
#> rho           0.1873  0.18174 0.004064       0.004064
#> sige        103.8701 23.29384 0.520866       0.566802
#> 
#> 2. Quantiles for each variable:
#> 
#>                2.5%       25%      50%      75%     97.5%
#> (Intercept) 30.5093 47.769573  56.4789  65.0877  84.86254
#> DISCBD      -9.2416 -6.153288  -4.7453  -3.3591  -0.67219
#> INC         -1.6655 -1.174948  -0.9451  -0.7102  -0.28365
#> HOVAL       -0.3744 -0.247270  -0.1775  -0.1128   0.01524
#> lag.INC     -0.8028  0.004752   0.4224   0.8433   1.66618
#> rho         -0.2116  0.072536   0.1836   0.3067   0.53579
#> sige        67.8946 87.054945 100.9815 116.4528 161.25826
#> 

print(summary(impacts(COL.D1.Bayes, tr=trMatc), short=TRUE, zstats=TRUE))

#> Impact measures (mixed, trace):
#>            Direct    Indirect      Total
#> DISCBD -4.8148501 -1.06067837 -5.8755285
#> INC    -0.9381442  0.29454370 -0.6436005
#> HOVAL  -0.1813164 -0.03994275 -0.2212591
#> ========================================================
#> MCMC sample impact results:
#> ========================================================
#> Simulated standard errors
#>           Direct   Indirect     Total
#> DISCBD 2.1926383 1.82145831 3.3798560
#> INC    0.3502903 0.81693522 0.8950370
#> HOVAL  0.1015612 0.06692703 0.1429623
#> 
#> Simulated z-values:
#>           Direct   Indirect      Total
#> DISCBD -2.219068 -0.7449248 -1.8410439
#> INC    -2.696796  0.3333552 -0.7511779
#> HOVAL  -1.802805 -0.7412143 -1.6277185
#> 
#> Simulated p-values:
#>        Direct   Indirect Total   
#> DISCBD 0.026482 0.45632  0.065615
#> INC    0.007001 0.73887  0.452546
#> HOVAL  0.071419 0.45856  0.103585

data(elect80, package="spData")
lw <- spdep::nb2listw(e80_queen, zero.policy=TRUE)
el_ml <- lagsarlm(log(pc_turnout) ~ log(pc_college) + log(pc_homeownership)
 + log(pc_income), data=elect80, listw=lw, zero.policy=TRUE, method="LU")
print(summary(el_ml))

#> 
#> Call:
#> lagsarlm(formula = log(pc_turnout) ~ log(pc_college) + log(pc_homeownership) + 
#>     log(pc_income), data = elect80, listw = lw, method = "LU", 
#>     zero.policy = TRUE)
#> 
#> Residuals:
#>        Min         1Q     Median         3Q        Max 
#> -1.1132095 -0.0690070 -0.0037755  0.0670525  0.8084858 
#> 
#> Type: lag 
#> Regions with no neighbours included:
#>  1183 1189 1832 2945 
#> Coefficients: (numerical Hessian approximate standard errors) 
#>                        Estimate Std. Error z value  Pr(>|z|)
#> (Intercept)            0.637925   0.041787 15.2660 < 2.2e-16
#> log(pc_college)        0.226366   0.015759 14.3645 < 2.2e-16
#> log(pc_homeownership)  0.481409   0.014524 33.1464 < 2.2e-16
#> log(pc_income)        -0.104942   0.016550 -6.3408 2.285e-10
#> 
#> Rho: 0.57742, LR test value: 1085.5, p-value: < 2.22e-16
#> Approximate (numerical Hessian) standard error: 0.014966
#>     z-value: 38.581, p-value: < 2.22e-16
#> Wald statistic: 1488.5, p-value: < 2.22e-16
#> 
#> Log likelihood: 2132.772 for lag model
#> ML residual variance (sigma squared): 0.013815, (sigma: 0.11754)
#> Number of observations: 3107 
#> Number of parameters estimated: 6 
#> AIC: -4253.5, (AIC for lm: -3170)
#> 

set.seed(1)
el_B <- spBreg_lag(log(pc_turnout) ~ log(pc_college) + log(pc_homeownership)
 + log(pc_income), data=elect80, listw=lw, zero.policy=TRUE)
print(summary(el_B))

#> 
#> Iterations = 501:2500
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>                           Mean        SD  Naive SE Time-series SE
#> (Intercept)            0.63963 0.0416231 9.307e-04      8.843e-04
#> log(pc_college)        0.22767 0.0159334 3.563e-04      3.563e-04
#> log(pc_homeownership)  0.48078 0.0149666 3.347e-04      3.347e-04
#> log(pc_income)        -0.10598 0.0165233 3.695e-04      3.695e-04
#> rho                    0.57602 0.0148031 3.310e-04      3.310e-04
#> sige                   0.01386 0.0003512 7.853e-06      7.177e-06
#> 
#> 2. Quantiles for each variable:
#> 
#>                          2.5%      25%      50%      75%    97.5%
#> (Intercept)            0.5589  0.61202  0.63883  0.66845  0.72085
#> log(pc_college)        0.1960  0.21724  0.22794  0.23789  0.25952
#> log(pc_homeownership)  0.4520  0.47081  0.48080  0.49069  0.51067
#> log(pc_income)        -0.1378 -0.11750 -0.10597 -0.09459 -0.07424
#> rho                    0.5478  0.56578  0.57579  0.58579  0.60480
#> sige                   0.0132  0.01362  0.01386  0.01408  0.01457
#> 

print(el_ml$timings)

#>           user.self elapsed
#> set_up        0.013   0.013
#> LU_set_up     0.003   0.003
#> LU_opt        0.215   0.219
#> coefs         0.002   0.002
#> LU_fdHess     0.314   0.315

print(attr(el_B, "timings"))

#>                   user.self sys.self elapsed user.child sys.child
#> set_up                0.010    0.000   0.010          0         0
#> SE_classic_set_up     3.084    0.012   3.105          0         0
#> complete_setup        0.000    0.000   0.000          0         0
#> sampling              1.268    0.001   1.274          0         0
#> finalise              0.017    0.000   0.017          0         0